% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_parallel.R
\name{activate_parallel}
\alias{activate_parallel}
\title{Internal function}
\usage{
activate_parallel(parallel = FALSE, NbeCores = 2)
}
\arguments{
\item{XY}{data.frame}

\item{mode}{character string either 'spheroid' or 'planar'. By default 'spheroid'}

\item{proj_type}{crs}
}
\description{
Activate paralle processing
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
\keyword{internal}
