% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.hull.poly.R
\name{alpha.hull.poly}
\alias{alpha.hull.poly}
\title{Internal function}
\usage{
alpha.hull.poly(
  XY,
  alpha = 1,
  buff = 0.1,
  exclude.area = FALSE,
  poly_exclude = NULL,
  mode = "spheroid",
  proj_type = "cea"
)
}
\arguments{
\item{XY}{data.frame coordinates}

\item{alpha}{integer if mode planar, in kilometers, if mode spheroid in
decimal degrees}

\item{buff}{numeric}

\item{exclude.area}{logical}

\item{poly_exclude}{sf polygon}

\item{mode}{character string either 'spheroid' or 'planar'. By default
'spheroid'}

\item{proj_type}{character string or numeric or object of crs class, by
default is "cea"}
}
\description{
Alpha hull process
}
\details{
The functions ahull_to_SPLDF and alpha.hull.poly were originally
posted in the website
https://casoilresource.lawr.ucdavis.edu/software/r-advanced-statistical-package/working-spatial-data/converting-alpha-shapes-sp-objects/
in a now broken link. It is also used in functions written by David Bucklin,
see https://github.com/dnbucklin/r_movement_homerange
}
