% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO.decline.R
\name{AOO.decline}
\alias{AOO.decline}
\title{Area of occupancy decline}
\usage{
AOO.decline(
  XY,
  Cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  proj_type = "cea",
  hab.map = NULL,
  hab.class = NULL,
  threat.map = NULL,
  all_individual_layers = FALSE
)
}
\arguments{
\item{XY}{\code{"dataframe"} see Details}

\item{Cell_size_AOO}{numeric, value indicating the grid size in kilometers used for estimating Area of Occupancy.  By default, equal to 2 km (i.e. 4 km2 grid cells)}

\item{nbe.rep.rast.AOO}{numeric , indicate the number of raster with random starting position for estimating the AOO. By default, it is 0 but some minimal translation of the raster are still done}

\item{parallel}{logical, whether running in parallel. By default, it is FALSE}

\item{NbeCores}{string integer, register the number of cores for parallel execution. By default, it is 2}

\item{show_progress}{logical, whether a bar showing progress in computation should be shown. By default, it is TRUE}

\item{proj_type}{character string or numeric or object of CRS class, by default is "cea"}

\item{hab.map}{raster, raster layer/stack or spatial polygons containing the
habitat spatial information}

\item{hab.class}{classes of values in \code{hab.map} to be considered as suitable}

\item{threat.map}{raster, raster layer/stack or spatial polygons (simple feature) indicating land cover unsuitable to taxa}

\item{all_individual_layers}{logical, by default FALSE, if TRUE compute AOO decline for each individual hab.map and threat.map provided}
}
\value{
\enumerate{
\item AOOs a data.frame of AOO estimates for each taxa and for each layer/spatial polygons if all_individual_layers is TRUE
\item AOO_decline a data.frame of AOO.decline in percentages
\item categories categories based on the sub-criteria of IUCN criterion A
}
}
\description{
Estimate areas of occupancy (AOO) decline for multiple taxa in square kilometers and percentage
`r lifecycle::badge("experimental")
}
\details{
\strong{Input} as a \code{dataframe} should have the following structure:

\strong{It is mandatory to respect field positions, but field names do not matter}

\tabular{ccc}{
\link{,1} \tab ddlat \tab numeric, latitude (in decimal degrees)\cr
\link{,2} \tab ddlon \tab numeric, longitude (in decimal degrees)\cr
\link{,3} \tab tax \tab character or factor, taxa names\cr
}

The argument \code{nbe.rep.rast.AOO} should ideally be higher than 20 for increasing
the chance to get the minimal number of occupied cell. Increasing \code{nbe.rep.rast.AOO} however
also increase the computing time. So this is a trade-off that depends on the importance to
get the minimal AOO and the size of the dataset.
}
\examples{
\dontrun{
data(land_cover_ex)
AOOs <- AOO.decline(XY = dataset.ex, hab.map = land_cover_ex, hab.class = c(40, 40))
}

}
\references{
Gaston & Fuller 2009 The sizes of species'geographic ranges, Journal of Applied Ecology, 49 1-9
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
