% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.legend.pos.R
\name{auto.legend.pos}
\alias{auto.legend.pos}
\title{Set Legend Position Automatically}
\usage{
auto.legend.pos(x, y, xlim = range(x), ylim = range(y))
}
\arguments{
\item{x}{a vector containing the (estimated) number of mature individuals of
the species}

\item{y}{a vector containing the years for which the number of mature
individuals was estimated}

\item{xlim}{a vector containing the names of the models to be fitted to
species population data}

\item{ylim}{a vector containing the years for which the number of mature
individuals should be predicted}
}
\description{
Set Legend Position Automatically
}
\references{
https://stackoverflow.com/questions/7198178/automatically-determine-position-of-plot-legend
}
\author{
user 'chan1142' in stackoverflow
}
