% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_basic.R
\name{bootstrap_basic}
\alias{bootstrap_basic}
\title{Basic bootstrap}
\usage{
bootstrap_basic(df, tmt, effect, cost, periods = NULL, i)
}
\arguments{
\item{df}{dataframe}

\item{tmt}{treatment arm identifier - currently only assuming two arms}

\item{effect}{effect column(s)}

\item{cost}{cost column (not multiple currently)}

\item{periods}{vector of weeks between each QoL measure, if applicable}

\item{i}{needed for \code{boot::boot}}
}
\value{
A named vector of mean incremental effect and cost and icer.
}
\description{
Basic bootstrap
}
\examples{
\dontshow{if (requireNamespace("boot")) withAutoprint(\{ # examplesIf}
df <- fake_health_ec_data
str(df)
# pass named args of bootstrap_basic to boot::boot
boot <- boot::boot(
  df,
  bootstrap_basic,
  R = 100,
  strata = df$tmt,
  tmt = "tmt",
  effect = dplyr::starts_with("utility"),
  cost = "cost",
  periods = c(10, 5, 20)
)
\dontshow{\}) # examplesIf}
}
