% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ind_qaly.R
\name{calc_ind_qaly}
\alias{calc_ind_qaly}
\title{Calculate an Individual QALY From a List of Utilities}
\usage{
calc_ind_qaly(utilities = c(), periods = c())
}
\arguments{
\item{utilities}{Utility scores, roughly between zero and one, in a vector.}

\item{periods}{Length of periods between utility scores, in weeks. For
example, utility may be measured at baseline, 6 weeks later, and 10 weeks
after baseline. Periods would then be c(6, 4). Periods is a vector one
element shorter than utilities.}
}
\value{
QALY, a numeric.
}
\description{
Calculate quality-adjusted life-years using utility scores and
the trapezium method for a single case.
}
\examples{
calc_ind_qaly(c(1, 0.4, 0.8), c(6, 4))
}
