% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_qaly.R
\name{calc_qaly}
\alias{calc_qaly}
\title{Calculate QALYs for a Data Frame}
\usage{
calc_qaly(df, qol, periods)
}
\arguments{
\item{df}{A dataframe.}

\item{qol}{Columns which contain QALY weights.}

\item{periods}{Vector containing the number of weeks between each measure.
The length of periods should be 1 less than the number of columns.}
}
\value{
vector of qaly values.
}
\description{
Calculates quality-adjusted life-years for each row of the dataframe.
Missing data will cause issues.
Note that a baseline differences between treatment arms should be accounted
for.
}
\examples{
df <- data.frame(
  id = 1:3,
  u0 = c(0, 0, 1),
  u1 = c(0.8, 0.4, 1),
  u2 = c(0.9, 0.7, 1),
  u3 = c(0.4, 0.6, 1)
)

calc_qaly(df, qol = dplyr::starts_with("u"), periods = c(4, 13, 26))
df |>
  dplyr::mutate(
    qaly = calc_qaly(df, qol = dplyr::starts_with("u"), periods = c(4, 13, 26))
  )
}
