% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_criterion_a.R
\name{cat_criterion_a}
\alias{cat_criterion_a}
\title{Categorize taxa according to IUCN criterion A}
\usage{
cat_criterion_a(
  A1_val = NULL,
  A2_val = NULL,
  A3_val = NULL,
  A4_val = NULL,
  A1.threshold = c(50, 70, 90),
  A234.threshold = c(30, 50, 80)
)
}
\arguments{
\item{A1_val}{numeric vector}

\item{A2_val}{numeric vector}

\item{A3_val}{numeric vector}

\item{A4_val}{numeric vector}

\item{A1.threshold}{numeric vector}

\item{A234.threshold}{numeric vector}
}
\value{
list
}
\description{
Provide IUCN category based on the four sub-criteria and provided thresholds
}
\details{
The function ...
}
\examples{
A1_val <- c(40, 90, 75, 10)
A2_val <- c(90, 95, 30, 10)
A3_val <- c(10, 30, 45, 15)
A4_val <- c(32, 18, 80, 20)
cat_criterion_a(
  A1_val = A1_val,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val
)

}
\references{

}
\author{
Dauby, G. & Lima, R.A.F.
}
