% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_criterion_b.R
\name{cat_criterion_b}
\alias{cat_criterion_b}
\title{Categorize taxa according to IUCN criterion B}
\usage{
cat_criterion_b(EOO = NULL, AOO = NULL, locations = NULL,
  protected = NULL, EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10), Loc.threshold = c(10, 5, 1))
}
\arguments{
\item{EOO}{numeric vector}

\item{AOO}{numeric vector}

\item{locations}{numeric vector}

\item{protected}{numeric vector}

\item{EOO.threshold}{numeric vector}

\item{AOO.threshold}{numeric vector}

\item{Loc.threshold}{numeric vector}
}
\value{
list
}
\description{
Provide IUCN category based on the four sub-criteria and provided thresholds
}
\details{
The function ...
}
\examples{

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(9, 12)
cat_criterion_b(EOO = EOO,AOO = AOO, locations = locations)

EOO <- c(34000, 80)
AOO <- c(300, 25)
locations <- c(1, 1)
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations)


}
\references{

}
\author{
Dauby, G. & Lima, R.A.F.
}
