% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_criterion_c.R
\name{cat_criterion_c}
\alias{cat_criterion_c}
\title{Categorize taxa according to IUCN criterion C}
\usage{
cat_criterion_c(
  C1_df = NULL,
  C2_df = NULL,
  C.threshold = c(10000, 2500, 250),
  C1.threshold = c(10, 20, 25),
  C2ai.threshold = c(1000, 250, 50),
  C2aii.threshold = c(90, 95, 100),
  mag.fluct = 10,
  high.alter = 80,
  all.cats = TRUE
)
}
\arguments{
\item{C1_df}{data frame with the parameters necessary to assess IUCN
sub-criterion C1.}

\item{C2_df}{data frame with the parameters necessary to assess IUCN
sub-criteria C2.}

\item{C.threshold}{numeric vector with the criterion C thresholds to define
small population sizes (e.g. number of mature individuals). Default values
are the thresholds recommended by the IUCN.}

\item{C1.threshold}{numeric vector with the C1 thresholds to convert
continuing decline estimates into categories. Default is the thresholds
recommended by IUCN.}

\item{C2ai.threshold}{numeric vector with the C2ai thresholds to assess the
number of mature individuals in each subpopulation. Default is the
thresholds recommended by IUCN.}

\item{C2aii.threshold}{numeric vector with the C2aii thresholds to assess the
proportion of mature individuals in one subpopulation. Default is the
thresholds recommended by IUCN.}

\item{mag.fluct}{numerical. Threshold of mean order of magnitude of the
differences between population minima and maxima to classify populations
with extreme fluctuations. Default to 10 as recommended by IUCN (2019).}

\item{high.alter}{numerical. Threshold of proportion of changes that are
followed by a change in the opposite direction. Default to 80\%, but
currently not implemented}

\item{all.cats}{logical. Should the categories from all criteria be returned
and not just the consensus categories? Default to TRUE.}
}
\value{
A list containing a vector of the consensus category from all
sub-criteria evaluated for each taxon (\code{ranks_C}) and the sub-criteria used
to obtain the consensus category (\code{cat_codes}). If \code{all.cats  == TRUE} the
function also returns a data frame containing the categories classified by
each sub-criteria individually (\code{all.cats}).
}
\description{
Provide the consensus IUCN category based on the sub-criteria of
IUCN criterion C (C1, C2ai, C2aii and C2b) and the thresholds recommended
by IUCN.
}
\details{
By default, the function provides the consensus category, following
the recommendations of IUCN (2019) that states "Only the criteria for the
highest category of threat that the taxon qualifies for should be listed".
Therefore, the consensus category is the highest category of threat among
the sub-criteria evaluated.

The function assumes that the order of the values in C1_df and C2_df are
from the same taxa (i.e. first element from C1_df and C2_df is always the
same species i). Therefore, the order of the estimates of population size
and continuing decline for each sub-criterion \emph{must} be the same.
}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and
Criteria. Version 14. Standards and Petitions Committee. Downloadable from:
http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Lima, R.A.F. & Dauby, G.
}
