% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_downlist.R
\name{cat_downlist}
\alias{cat_downlist}
\title{Downlist Threat Categories}
\usage{
cat_downlist(cats = NULL, down.by = 1, dd = FALSE, sign = TRUE)
}
\arguments{
\item{cats}{character. The vector containing the IUCN threat categories.}

\item{down.by}{numerical. The number of steps to downlist the categories.
Default to 1.}

\item{dd}{logical. Should the Data Deficient (DD) category be included in the
downlisting? Defaut to FALSE.}

\item{sign}{logical. Should the degrees sign be indicated in the downlisted
category. Default to TRUE.}
}
\description{
Perform the downlisting of threat categories which is often
  necessary for regional conservation assessmentes (IUCN 2012).
}
\details{
For regional conservation assessments, the IUCN recommends an
  additional step, which relates to the possible "effect of populations of
  the same taxon in neighbouring regions on the regional population", and
  thus to the possibility of a possible rescue effect (IUCN 2012).

  Although the IUCN (2012) considers the exceptional possibility of
  uplisting, the function only performs downliting of the categories of
  threat. Downlisting normally is a one-step change in the category (i.e.
  from EN to VU), but this can be controlled by the argument ```down.by```
  (default to 1).

  By default, the Data Deficient category is excluded from the donwlisting.
  The Least Concern category remains unaltered as well.

  Note that "if it is unknown whether or not extra-regional populations
  influence the extinction risk of the regional population, the category
  (...) should be kept unaltered" (IUCN 2012).
}
\examples{
cats <- c("EX","CR","EN","VU","NT","DD","LC","NA","NE")
cat_downlist(cats)
cat_downlist(cats, down.by = 2)
cat_downlist(cats, dd = TRUE)
cat_downlist(cats, sign = FALSE)
cat_downlist(cats, down.by = 2, dd = TRUE, sign = FALSE)



}
\references{
IUCN (2012). Guidelines for Application of IUCN Red List Criteria
  at Regional and National Levels (Version 4.0). IUCN. Gland, Switzerland and
  Cambridge, UK. 41pp.
}
\author{
Lima, R.A.F.
}
