% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.occupied.R
\name{cell.occupied}
\alias{cell.occupied}
\title{Internal function}
\usage{
cell.occupied(
  nbe_rep = 0,
  size = 4,
  coord,
  export_shp = TRUE,
  proj_type = NULL
)
}
\arguments{
\item{nbe_rep}{integer}

\item{size}{integer}

\item{coord}{data.frame}

\item{export_shp}{logical}
}
\description{
Count number of occupied cells given resolution, projection
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
