% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conr-package.R
\docType{package}
\name{conr-package}
\alias{conr}
\alias{conr-package}
\title{conr: conr: Helping with Health Economics Work}
\description{
conr is a package to provide templates and convenience functions for work in health economics. It is my first package and will be bug-ridden.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/connor-ballinger/conr}
  \item Report bugs at \url{https://github.com/connor-ballinger/conr/issues}
}

}
\author{
\strong{Maintainer}: Connor Ballinger \email{cdballinger97@gmail.com} (\href{https://orcid.org/0009-0004-5977-5385}{ORCID})

}
\keyword{internal}
