% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convex.Hull.Poly.R
\name{Convex.Hull.Poly}
\alias{Convex.Hull.Poly}
\title{Internal function}
\usage{
Convex.Hull.Poly(
  XY,
  mode = "spheroid",
  exclude.area = FALSE,
  poly_exclude = NULL,
  proj_type = "cea"
)
}
\arguments{
\item{XY}{data.frame}

\item{mode}{character string either 'spheroid' or 'planar'. By default 'spheroid'}

\item{exclude.area}{logical}

\item{poly_exclude}{polygon}

\item{proj_type}{character string or numeric or object of CRS class, by default is "cea"}
}
\description{
Build convex hull polygon
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
