% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.check.R
\name{coord.check}
\alias{coord.check}
\title{Coordinates check}
\usage{
coord.check(XY, listing = TRUE, proj_type = NULL, listing_by_valid = FALSE)
}
\arguments{
\item{XY}{}

\item{listing}{}

\item{proj_type}{}
}
\value{

}
\description{
Coordinates check
}
