% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_B.R
\name{criterion_B}
\alias{criterion_B}
\title{Assess IUCN Criterion B}
\usage{
criterion_B(
  x,
  protec.areas = NULL,
  EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10),
  Loc.threshold = c(10, 5, 1),
  SubPop = TRUE,
  Resol_sub_pop = 5,
  Cell_size_locations = 10,
  method_locations = "fixed_grid",
  Rel_cell_size = 0.05,
  method_protected_area = "no_more_than_one",
  ID_shape_PA = "WDPA_PID",
  country_map = NULL,
  method.range = "convex.hull",
  alpha = 1,
  buff.alpha = 0.1,
  exclude.area = FALSE,
  buff_width = 0.1,
  Cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2,
  proj_type = "cea",
  mode = "spheroid",
  DrawMap = FALSE
)
}
\arguments{
\item{x}{a \code{dataframe} or an object of class \code{spgeoIN} see
\url{https://github.com/azizka/speciesgeocodeR}. See Details}

\item{protec.areas}{a \code{SpatialPolygonsDataFrame}, shapefile with
protected areas.  If provided, this will be taken into account for
calculating number of location (see Details and
\code{method_protected_area}).  By default, no shapefile is provided}

\item{EOO.threshold}{numeric vector indicating the thresholds used to categorize EOO in IUCN categories}

\item{AOO.threshold}{numeric vector indicating the thresholds used to categorize AOO in IUCN categories}

\item{Loc.threshold}{numeric vector indicating the thresholds used to categorize the number of locations in IUCN categories}

\item{SubPop}{logical if the number of sub-populations should be computed. By default is TRUE}

\item{Resol_sub_pop}{a numeric, value indicating the radius size in
kilometers used for estimating the number of sub-population. By default,
equal to 5}

\item{Cell_size_locations}{a numeric, value indicating the grid size in
kilometers used for estimating the number of location. By default, equal to
10}

\item{method_locations}{a character string, indicating the method used for
estimating the number of locations.  "fixed_grid" or "sliding scale". See
details. By default, it is "fixed_grid"}

\item{Rel_cell_size}{a numeric, if \code{method_locations="sliding scale"},
\code{Cell_size_locations} is ignored and the resolution is given by the
maximum distance separating two occurrences multiplied by
\code{Rel_cell_size}. By default, it is 0.05}

\item{method_protected_area}{a character string. By default is
"no_more_than_one"", which means occurrences within protected areas (if
provided) will not be taken into account for estimating the number of
locations following the grid system, see Details. By default, it is
"no_more_than_one"}

\item{ID_shape_PA}{a character string, indicating the field name of
\code{protec.areas} with ID of the \code{SpatialPolygonsDataFrame} of
protected areas}
}
\value{

}
\description{
Preliminary assessment of species conservation status following
 IUCN Criterion B, which is based on species geographic distribution (i.e. extent
 of occurrence - EOO, and area of occupancy, AOO)
}
\details{
The function ...
}
\examples{

}
\references{

}
\author{
Dauby, G. & Lima, R.A.F.
}
