% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_C.R
\name{criterion_C}
\alias{criterion_C}
\title{Assess IUCN Criterion C}
\usage{
criterion_C(x)
}
\arguments{
\item{x}{the character string.}
}
\value{

}
\description{
Preliminary assessment of species conservation status following
 IUCN Criterion C, which is based on small population size and decline.
}
\details{
The function ...
}
\examples{


}
\references{

}
\author{
Lima, R.A.F. & Dauby, G.
}
