% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_D.R
\name{criterion_D}
\alias{criterion_D}
\title{Assess IUCN Criterion D}
\usage{
criterion_D(x)
}
\arguments{
\item{x}{the character string.}
}
\value{

}
\description{
Preliminary assessment of species conservation status following
 IUCN Criterion D, which is based on very small or restricted population 
 (Criteria D1 and D2).
}
\details{
The function ...
}
\examples{


}
\references{

}
\author{
Lima, R.A.F. & Dauby, G.
}
