% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantspeciesdistribution.R
\name{dataset.ex}
\alias{dataset.ex}
\title{Dataset of plant species distribution}
\format{A dataframe}
\usage{
data(dataset.ex)
}
\description{
A dataframe containing georeferenced occurrences
 in decimal degrees of several plant species, organized as below
}
\details{
\itemize{
  \item ddlat. latitude in decimal degrees
  \item ddlon. longitude in decimal degrees
  \item tax. species name
}
}

