% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_progess_bar.R
\name{display_progress_bar}
\alias{display_progress_bar}
\title{Internal function}
\usage{
display_progress_bar(show_progress = TRUE, max_pb)
}
\arguments{
\item{show_progress}{logical}

\item{max_pb}{integer}
}
\description{
Display progress
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
\keyword{internal}
