% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.ahull_to_SPLDF}
\alias{.ahull_to_SPLDF}
\title{Internal function}
\usage{
.ahull_to_SPLDF(x, proj4string = NA)
}
\description{
Alpha hull processing
}
\details{
The functions ahull_to_SPLDF and alpha.hull.poly were originally posted in the website https://casoilresource.lawr.ucdavis.edu/software/r-advanced-statistical-package/working-spatial-data/converting-alpha-shapes-sp-objects/
in a now broken link. It is also used in functions written by David Bucklin, see https://github.com/dnbucklin/r_movement_homerange
}
