% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.alpha.hull.poly}
\alias{.alpha.hull.poly}
\title{Internal function}
\usage{
.alpha.hull.poly(XY, alpha = 1, buff = 0.1)
}
\description{
Alpha hull process
}
\details{
The functions ahull_to_SPLDF and alpha.hull.poly were originally posted in the website https://casoilresource.lawr.ucdavis.edu/software/r-advanced-statistical-package/working-spatial-data/converting-alpha-shapes-sp-objects/
in a now broken link. It is also used in functions written by David Bucklin, see https://github.com/dnbucklin/r_movement_homerange
}
