% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.AOO.estimation}
\alias{.AOO.estimation}
\title{Internal function}
\usage{
.AOO.estimation(coordEAC, cell_size = 2, nbe_rep = 0, export_shp = FALSE)
}
\arguments{
\item{coordEAC}{data.frame}

\item{cell_size}{integer}

\item{nbe_rep}{integer}

\item{export_shp}{logical}
}
\description{
AOO estimatiion
}
