% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.Convex.Hull.Poly}
\alias{.Convex.Hull.Poly}
\title{Internal function}
\usage{
.Convex.Hull.Poly(XY)
}
\description{
Build convex hull polygon
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
