% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.comp.R
\name{.generate_loc}
\alias{.generate_loc}
\title{Internal function.}
\usage{
.generate_loc(
  dataset,
  method = "fixed_grid",
  nbe_rep = 0,
  Cell_size_locations = 10,
  Rel_cell_size = 0.05,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  proj_type = "cea"
)
}
\arguments{
\item{dataset}{list.}
}
\description{
Internal function.
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
