% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over.valid.poly.R
\name{.over.valid.poly}
\alias{.over.valid.poly}
\title{Internal function}
\usage{
.over.valid.poly(poly, points, proj_user = NULL, min.dist = 0.1,
  value = "dist")
}
\arguments{
\item{poly}{Spatial polygon}

\item{points}{XY data frame}

\item{min.dist}{minimum tolerated distance between polygons and points.
Default to 0.1 m.}

\item{value}{output value: proportional distance ("dist") or inside/outside
the polygon ("flag")?}

\item{proj_type}{projected coordinate system (in meters)}
}
\description{
Get Occurrences Distance to a Polygon
}
\details{
The spatial polygon must be a `SpatialPolygonsDataFrame` in whhich
  each polygon/feature is one taxon, an the data frame contains a column
  `tax` with the taxa name. The XY data frame has the same structure as other
  XY objects within `ConR` with the three first columns being `ddlat`,
  `ddlon` and `tax`, but also a column `classes` with the classes of
  conficence level, which is represented by numbers in increasing order of
  confidence. For instance, if confidence levels are "low" and "high", the
  corresponding values in columns `classes` should be 1 and 2.
}
\examples{

mydf <- data.frame(ddlat = c(-44.6,-46.2,-45.4,-42.2,-43.7,-45.0,-28.0),
                   ddlon = c(-42.2,-42.6,-45.3,-42.5,-42.3,-39.0,-17.2),
                   tax = rep("a", 7),
                   valid = c(c(TRUE,TRUE,TRUE,FALSE,TRUE,TRUE,FALSE)),
                   stringsAsFactors = FALSE)
mydf$classes = as.double(mydf$valid)
shp <- SpatialPolygonsDataFrame(EOO.computing(mydf[mydf$valid,], export_shp = TRUE)[[2]], data.frame(tax = "a"))
plot(mydf[,2:1])
plot(shp, add=TRUE)
.over.valid.poly(shp, mydf)
.over.valid.poly(shp, mydf, proj_user = 5641)
.over.valid.poly(shp, mydf, proj_user = 5641, value = "flag")  

}
