% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.prop_threat}
\alias{.prop_threat}
\title{Internal function}
\usage{
.prop_threat(Cell_count, threshold)
}
\arguments{
\item{Cell_count}{data.frame}

\item{threshold}{integer}
}
\description{
Compute prop and nbr taxa per cell
}
