% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.subpop.comp}
\alias{.subpop.comp}
\title{Internal function}
\usage{
.subpop.comp(XY, Resol_sub_pop)
}
\arguments{
\item{XY}{data.frame}

\item{Resol_sub_pop}{integer}
}
\description{
subpopulations estimatiion
}
