% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_B.R
\name{draw_map_cb}
\alias{draw_map_cb}
\title{Internal function}
\usage{
draw_map_cb(
  XY,
  name_sp,
  eoo_poly,
  aoo_poly,
  locations_poly,
  subpop,
  proj_type,
  results,
  add.legend = TRUE
)
}
\arguments{
\item{XY}{data.frame, of at least two columns (coordinates), third is taxa}

\item{name_sp}{character vector}

\item{eoo_poly}{sf}

\item{aoo_poly}{sf}

\item{locations_poly}{sf}

\item{subpop}{sf}

\item{proj_type}{crs}

\item{results}{data.frame}

\item{add.legend}{logical}
}
\value{
a plot
}
\description{
Coordinates check
}
\keyword{internal}
