% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO.comp.R
\name{EOO.comp}
\alias{EOO.comp}
\title{Compute Extent of Occurrence}
\usage{
EOO.comp(
  XY,
  exclude.area = FALSE,
  buff_width = 0.1,
  country_map = NULL,
  Name_Sp = "tax",
  method.range = "convex.hull",
  alpha = 1,
  buff.alpha = 0.1,
  method.less.than3 = "not comp",
  mode = "spheroid",
  proj_type = "cea"
)
}
\arguments{
\item{XY}{data.frame}

\item{exclude.area}{logical. Default if FALSE}

\item{buff_width}{numeric. Default if 0.1}

\item{country_map}{SpatialPolygonDataframe. Default if NULL}

\item{Name_Sp}{string}

\item{method.range}{string}

\item{alpha}{integer}

\item{buff.alpha}{numeric}

\item{method.less.than3}{string}

\item{mode}{character string either 'spheroid' or 'planar'. By default 'spheroid'}

\item{proj_type}{character string or numeric or object of CRS class, by default is "cea"}
}
\value{
A list with one element being a numeric vector [[1]]EOO and the spatial polygon [[2]]spatial.polygon
}
\description{
Compute EOO given a data frame of coordinate in decimals degrees
}
\details{
If `exclude.area` is TRUE and country_map is not provided, 
the world country polygons used comes from the package \href{https://www.rdocumentation.org/packages/rnaturalearth/versions/0.1.0/topics/ne_countries}{rnaturalearth}

By default (`mode = "spheroid"`),the area of the polygon is based on a polygon in longitude/latitude coordinates considering the earth as an ellipsoid. 
See \href{https://CRAN.R-project.org/package=geosphere}{geosphere} package.
To make a polygon more accurate, the function use the function `makePoly` from the \href{https://CRAN.R-project.org/package=geosphere}{geosphere} package.
This adds vertices on the great circles (see example below) and can make important difference for species with large distribution.

An estimation of EOO based on projected data is also possible (`mode = "planar"`).
This allow the user to use its own projection.
By default, the projection (`proj_type = "cea"`) is a \href{https://epsg.io/6933}{global cylindrical equal-area projection}.
It can makes sense to use a planar projection to estimate the EOO. See example.

It is possible to use another projection by providing the EPSG code to `proj_type`. See example.
Check \href{https://www.rdocumentation.org/packages/rgdal/versions/0.4-10/topics/make_EPSG}{make_EPSG}
from the `rgdal` packag for getting the EPSG code.
}
\examples{

country_map <-
  rnaturalearth::ne_countries(scale = 50, returnclass = "sp")

### example Large distribution
# Spheroid estimation
XY <- rbind(c(-180,-20), c(-160,5), c(-60, 0), c(-160,-60), c(-180,-20))
p1 <- 
  EOO.comp(XY =  XY[,c(2, 1)])
plot(country_map)
plot(p1$spatial.polygon, lwd = 2, col = 'red', add = TRUE)

p2 <- 
  EOO.comp(XY = XY[,c(2, 1)], mode = "planar")
plot(st_as_sf(country_map), reset = FALSE, extent = as(p2$spatial.polygon, "sf"))
plot(p2$spatial.polygon, lwd = 2, col = 'red', add = TRUE)

# World Mercartor projection
p2 <- 
  EOO.comp(XY = XY[,c(2, 1)], mode = "planar", proj_type = 3395)
plot(st_as_sf(country_map), reset = FALSE, extent = as(p2$spatial.polygon, "sf"))
plot(p2$spatial.polygon, lwd = 2, col = 'red', add = TRUE)

### example Antartic distribution
XY <- rbind(c(-150, -65), c(0, -62), c(120, -78), c(150, -65))

p1 <- 
  EOO.comp(XY = XY[,c(2, 1)], mode = "planar", proj_type = "Antarctic")

p1 <- st_transform(as(p1$spatial.polygon, "sf"), proj_crs(proj_type = "Antarctic"))
plot(st_transform(st_as_sf(country_map), proj_crs(proj_type = "Antarctic")), extent = p1)
plot(p1, lwd = 2, col = 'red', add = TRUE)

p1 <- 
  EOO.comp(XY = XY[,c(2, 1)])

p1 <- st_transform(as(p1$spatial.polygon, "sf"), proj_crs(proj_type = "Antarctic"))
plot(st_transform(st_as_sf(country_map), proj_crs(proj_type = "Antarctic")), extent = p1)
plot(p1, lwd = 2, col = 'red', add = TRUE)

}
\author{
Dauby, G. & Lima, R.A.F.
}
