% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.ex.R
\name{example_criterionA}
\alias{example_criterionA}
\title{Dataset to illustrate the assessment of IUCN criterion A}
\format{
A data frame with six rows and 33 variables
}
\source{
\url{https://www.iucnredlist.org/resources/criterion-a}
}
\usage{
data(example_criterionA)
}
\description{
A dataset containing the population sizes of six taxa and from 1970 until
2032 at every two years. First two species are taken from the same example
used by IUCN (2019) to illustrate the assessment of this criterion.
}
\keyword{datasets}
