% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.ex.R
\name{example_criterionC}
\alias{example_criterionC}
\title{Dataset to illustrate the assessment of IUCN criterion C}
\format{
A data frame with 9 rows and 7 variables
}
\usage{
data(example_criterionC)
}
\description{
A dataset containing the population sizes for nine
different taxa and from 1970 until 2000 at every five years,
without information on subpopulation sizes.
}
\keyword{datasets}
