% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.ex.R
\name{example_fluctuation}
\alias{example_fluctuation}
\title{Dataset to illustrate the assessment of populations with extreme fluctuation}
\format{
A data frame with seven rows and 28 variables
}
\source{
\url{http://www.iucnredlist.org/documents/RedListGuidelines.pdf}
}
\usage{
data(example_fluctuation)
}
\description{
A dataset containing the population sizes adapted from some of the panels 
provided in Figure 4.4. of IUCN (2019) to illustrate different examples
of how population can fluctuate.
}
\keyword{datasets}
