% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.ex.R
\name{example_tutorial}
\alias{example_tutorial}
\title{Dataset to run the package introductory vignette}
\format{
A named list with two data frames: 'occurrences' and 'population.sizes'
}
\source{
\url{https://doi.org/10.15468/dl.mzmat2}
}
\usage{
data(example_tutorial)
}
\description{
A dataset containing occurrence records and population size estimates from
three arborescent species from the Atlantic Forest obtained from different
sources, but mostly from GBIF and TreeCo.
}
\details{
Only occurrence records considered to have a high confidence level in terms
of georefferencing and species identifications are included.

In data frame 'occurrences' columns are: ddlat, ddlon, tax, coly, UC, source

In data frame 'population.sizes' columns are: species, and all years with
population estimates
}
\keyword{datasets}
