% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_health_ec_data}
\alias{fake_health_ec_data}
\title{Some fake data I made}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 11 columns.
}
\usage{
fake_health_ec_data
}
\description{
Includes a step-wedge design - 5 clusters, 300-odd patients.
}
\keyword{datasets}
