% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_basic.R
\name{fn_calc_increment}
\alias{fn_calc_increment}
\title{Helper function for second row minus first row}
\usage{
fn_calc_increment(df, col)
}
\arguments{
\item{df}{dataframe}

\item{col}{variable}
}
\value{
vector created by \code{diff}.
}
\description{
Helper function for second row minus first row
}
