% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{Long Format Date}
\usage{
format_date(date = Sys.Date(), ...)
}
\arguments{
\item{date}{Currently a Date format, as provided by Sys.Date()}

\item{...}{Date.}
}
\value{
character date
}
\description{
Long Format Date
}
\examples{
format_date()
format_date(as.Date("2023-01-01"))
}
