% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_docx.R
\name{format_docx}
\alias{format_docx}
\title{Produce a Word Document}
\usage{
format_docx()
}
\value{
\code{officedown::rdocx_document} to pass to \code{rmarkdown::render}.
}
\description{
Provides a reference document with styles.
}
