% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.patches.R
\name{frag_assess}
\alias{frag_assess}
\title{Assessment of fragmentation subpopulations}
\usage{
frag_assess(
  XY,
  cell_size = 2,
  nbe_rep = 0,
  AOO = NULL,
  Resol_sub_pop = NULL,
  subpop_poly = NULL,
  dist_isolated = NULL,
  proj_type = "cea",
  export_shp = FALSE,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2
)
}
\arguments{
\item{XY}{data frame}

\item{cell_size}{numeric}

\item{nbe_rep}{numeric}

\item{AOO}{numeric. AOO of the species in squared kilometres.}

\item{Resol_sub_pop}{numeric. Defines in kilometres the radius of the circles
around each occurrence}

\item{subpop_poly}{Simple feature collection, output of
\code{subpop.estimation} function}

\item{dist_isolated}{numeric vector providing a value for each species. Distance in kilometres for identifying
subpopulations considered to be isolated}

\item{proj_type}{projected coordinate system (in meters)}

\item{export_shp}{logical}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Assess whether species populations are severaly fragmented
}
\details{
This function evaluates the proportion of the total area of
occupancy in habitat patches separated from others by a large distance. Based
on IUCN guidelines page(48): "A taxon can be considered to be severely
fragmented if most (>50\%) of its total area of occupancy is in habitat
patches that are (...) (2) separated from other habitat patches by a large
distance."

This function interpret subpopulations as obtained using
\code{subpop.estimation} function as habitat patches. First, subpopulations
that are isolated i.e. distant to all others subpopulations by at least
\code{dist_isolated} kilometres, are identified. Then, the percentage of the
AOO concerned by those "isolated" subpopulations is calculated.
}
\examples{

\donttest{
mydf <- data.frame(ddlat = c(-44.6,-46.2,-45.4,-42.2,-43.7,-45.0,-48.0),
                   ddlon = c(-42.2,-42.6,-45.3,-42.5,-42.3,-39.0,-37.2),
                   tax = rep("a", 7),
                   stringsAsFactors = FALSE)

get.patches(XY = mydf, dist_isolated = 200)
}



}
