% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severe_frag.R
\name{get.patches}
\alias{get.patches}
\title{Internal function}
\usage{
get.patches(
  subpop_poly = NULL,
  dist_isolated = NULL,
  proj_type = "cea",
  export_shp = FALSE,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2
)
}
\arguments{
\item{subpop_poly}{Simple feature collection, output of
\code{subpop.estimation} function}

\item{dist_isolated}{numeric. Distance in kilometres for identifying
subpopulations considered to be isolated}

\item{proj_type}{string or numeric}

\item{export_shp}{logical}

\item{parallel}{a logical. Whether running in parallel. By default, it is FALSE}

\item{NbeCores}{an integer. Register the number of cores for parallel execution. By default, it is 2}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get Occupied Cells and Patches
}
\keyword{internal}
