% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_format.R
\name{html_format}
\alias{html_format}
\title{Produce an HTML Report}
\usage{
html_format()
}
\value{
Not sure, essentially a html report.
}
\description{
Produce an HTML Report
}
\examples{
output: conr::html_format
}
