% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initiate project}
\usage{
init_project(path, use_renv = TRUE, use_git = TRUE, use_folders = TRUE, ...)
}
\arguments{
\item{path}{String for desired directory.}

\item{use_renv}{TRUE/FALSE Use renv?}

\item{use_git}{TRUE/FALSE Use Git?}

\item{use_folders}{TRUE/FALSE Use the default directory structure, with some
blank files.}

\item{...}{Dots.}
}
\value{
A project.
}
\description{
Create a new project with some folder structures and preferences
in place. Similar examples:

\url{https://rstudio.github.io/rstudio-extensions/rstudio_project_templates.html}

\url{https://github.com/jonas-hag/analysistemplates/}

\url{https://github.com/rostools/prodigenr/blob/main/R/setup_project.R}
}
