% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_heading.R
\name{insert_heading}
\alias{insert_heading}
\title{Insert New Heading}
\usage{
insert_heading()
}
\value{
Hyphens up to the 80 character limit.
}
\description{
Create a new heading for a script by inserting hyphens from
cursor position to the 80 character limit.
Function is linked to a .dcf file, thereby creating an Addin.
Create a keyboard shortcut (e.g. Ctrl + H) for the Addin.

Credit to \url{https://github.com/mnist91/shoRtcut}.
}
