% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCN.comp.R
\name{IUCN.comp}
\alias{IUCN.comp}
\title{Internal function}
\usage{
IUCN.comp(
  DATA,
  poly_borders = NULL,
  Cell_size_AOO = 2,
  Cell_size_locations = 10,
  Resol_sub_pop = 5,
  method_locations = c("fixed_grid"),
  Rel_cell_size = 0.05,
  protec.areas = NULL,
  exclude.area = FALSE,
  method_protected_area = "no_more_than_one",
  ID_shape_PA = "WDPA_PID",
  buff_width = 0.1,
  NamesSp = "species1",
  write_shp = FALSE,
  file_name = NULL,
  add.legend = TRUE,
  DrawMap = TRUE,
  map_pdf = FALSE,
  draw.poly.EOO = TRUE,
  SubPop = TRUE,
  MinMax,
  alpha = 1,
  buff.alpha = 0.1,
  method.range = "convex.hull",
  nbe.rep.rast.AOO = 0,
  verbose = TRUE,
  showWarnings = TRUE
)
}
\arguments{
\item{DATA}{data.frame}

\item{poly_borders}{SpatialPolygonDataFrame}

\item{Cell_size_AOO}{integer}

\item{Cell_size_locations}{integer}

\item{Resol_sub_pop}{integer}

\item{method_locations}{integer}

\item{Rel_cell_size}{integer}

\item{protec.areas}{SpatialPolygonDataFrame}

\item{exclude.area}{logical}

\item{method_protected_area}{string}

\item{ID_shape_PA}{string}

\item{buff_width}{numeric}

\item{NamesSp}{string}

\item{write_shp}{logical}

\item{file_name}{string}

\item{add.legend}{logical}

\item{DrawMap}{logical}

\item{map_pdf}{logical}

\item{draw.poly.EOO}{logical}

\item{SubPop}{logical}

\item{MinMax}{numeric vector}

\item{alpha}{integer}

\item{buff.alpha}{numeric}

\item{method.range}{string}

\item{nbe.rep.rast.AOO}{integer}

\item{verbose}{logical}

\item{showWarnings}{logical}
}
\description{
Compute IUCN eval
}
