% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_preferences.R
\name{keep_prefs}
\alias{keep_prefs}
\alias{copy_prefs}
\title{Keep/Copy preferences}
\usage{
keep_prefs(doc_id)

copy_prefs(doc_id, prefs, new_prefs)
}
\arguments{
\item{doc_id}{Open preferences file identifier.}

\item{prefs}{Existing preferences file.}

\item{new_prefs}{Preferences file from conr.}
}
\description{
Keep/Copy preferences
}
