% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_preferences.R
\name{keep_prof}
\alias{keep_prof}
\alias{copy_prof}
\title{Keep/Copy profile}
\usage{
keep_prof()

copy_prof(new, old)
}
\arguments{
\item{new}{New profile.}

\item{old}{Old profile.}
}
\description{
Keep/Copy profile
}
