% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_df.R
\name{knit_docx_df}
\alias{knit_docx_df}
\title{Knit method for dataframes in docx output}
\usage{
knit_docx_df(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{dots}
}
\value{
not sure... whatever \code{knitr::knit_print} returns.
}
\description{
Knit method for dataframes in docx output
}
