% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_df.R
\name{knit_html_df}
\alias{knit_html_df}
\title{Knit method for dataframes in html output}
\usage{
knit_html_df(
  df,
  DT_opts = list(rownames = TRUE, dom = "ltipB", filter = "top", scrollX = TRUE,
    pageLength = 5, lengthMenu = c(5, 10, 20), buttons = c("copy", "csv")),
  ...
)
}
\arguments{
\item{df}{Dataframe}

\item{DT_opts}{passed to \code{\link[DT]{datatable}}.}

\item{...}{Unused dots.}
}
\value{
Returns the dataframe in the form of \code{\link[DT]{datatable}}, via
\code{knitr::knit_print}.
}
\description{
Currently not ideal as \code{\link[DT]{datatable}} trims trailing zeros,
leading to inconsistent number of decimal places in a given column.
\code{\link[DT]{formatRound}} is not an option as trailing zeros are never
trimmed.
}
\details{
Why \code{DT} over \code{gt}? \code{DT}: pages, choose pageLength, copy/csv buttons. \code{gt}:
has most of this here
\url{https://gt.rstudio.com/reference/opt_interactive.html}, but not the
download option.
}
\seealso{
\code{\link[=knit_df]{knit_df()}}, \code{\link[=knit_docx_df]{knit_docx_df()}}
}
