% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{land}
\alias{land}
\title{land}
\format{
A dataframe
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}
\usage{
data(land)
}
\description{
A Simple feature collection with 1 feature MULTIPOLYGON showing land border WGS 84 CRS
}
\keyword{datasets}
