% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_cover_ex.R
\name{land_cover_ex}
\alias{land_cover_ex}
\title{Dataset of land cover changes for two years}
\format{
A RasterBrick
}
\source{
The European Space Agency
}
\usage{
data(land_cover_ex)
}
\description{
A RasterBrick of land cover change across a small area in Brazil
}
\references{
http://www.esa-landcover-cci.org/
}
\keyword{datasets}
