% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locations.estimation.R
\name{Locations.estimation}
\alias{Locations.estimation}
\title{Internal function}
\usage{
Locations.estimation(
  coordEAC,
  cell_size = 10,
  nbe_rep = 0,
  export_shp = FALSE,
  proj_type = proj_type,
  method = "fixed_grid",
  Rel_cell_size = 0.05
)
}
\arguments{
\item{coordEAC}{data.frame}

\item{cell_size}{integer}

\item{nbe_rep}{integer}

\item{export_shp}{logical}

\item{proj_type}{character string}

\item{Rel_cell_size}{numeric, if \code{method_locations="sliding scale"}, \code{Cell_size_locations} is ignored and the resolution is given by the maximum distance separating two occurrences multiplied by \code{Rel_cell_size}. By default, it is 0.05}
}
\description{
Number of location estimation
}
