% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Malagasy.amphibian.R
\name{Malagasy.amphibian}
\alias{Malagasy.amphibian}
\title{Dataset of Malagasy Amphibian occurrences}
\format{
A dataframe
}
\source{
GBIF.org (9th February 2016) GBIF Occurrence Download \url{http://doi.org/10.15468/dl.2tkoae}
}
\usage{
data(Malagasy.amphibian)
}
\description{
A SpatialPolygonDataFrame showing land cover
}
\details{
This dataset was further filtered out for keeping only species that have at least one occurrence in Madagascar
}
\references{
Download Information:
DOI: http://doi.org/10.15468/dl.2tkoae (may take some hours before being active)
Creation Date: Tuesday, February 9, 2016 9:31:26 PM CET
Records included: 74352 records from 50 published datasets
Data size: 4.6 MB
Download format: SIMPLE_CSV
Filter used: TaxonKey: Amphibia Linnaeus, 1758
Geometry: POLYGON((-17.578125 33.724339,-17.578125 -36.597889,56.953125 -36.597889,56.953125 33.724339,-17.578125 33.724339))
HasCoordinate: true
HasGeospatialIssue: false
BasisOfRecord: Human Observation or Observation or Living Specimen or Literature Occurrence or Specimen
}
\keyword{datasets}
