% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_preferences.R
\name{overwrite_preferences}
\alias{overwrite_preferences}
\title{Write preferences}
\usage{
overwrite_preferences()
}
\value{
Invisible; path to \code{"rstudio-prefs.json"}.
}
\description{
Existing preferences can be checked using
\code{\link[usethis]{edit_rstudio_prefs}}.
}
\details{
Function must be used interactively. A menu asks you to confirm the decision
to overwrite preferences.

To view the preferences file in the conr package, run the following code in
the console:

\code{
  system.file(package = "conr", "preferences", "rstudio-prefs.json") |>
    file.edit()
}

Some preferences you may wish to change:
\itemize{
\item "default_project_location": "~/coding",
\item "initial_working_directory": "~/coding",
\item "default_open_project_location": "~/coding",
\item "editor_theme": "Chrome" or "Tomorrow Night" or ...
}

Preferences can be most easily be changed using RStudio menus.

I was going to use \code{usethis:::rstudio_config_path} but apparently internal
functions should not be used outside of a given package.
}
\seealso{
\code{\link[=overwrite_profile]{overwrite_profile()}}
}
