% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_preferences.R
\name{overwrite_profile}
\alias{overwrite_profile}
\title{Write .Rprofile}
\usage{
overwrite_profile()
}
\value{
Invisible; path to \code{.Rprofile}.
}
\description{
This function overwrites the existing user \code{.Rprofile} with one found in
\code{conr}. The file includes some simple options documented in the \code{usethis}
package.
}
\details{
Note that the function only runs interactively and a mini menu is displayed
after running the function to confirm your choice. If you already have code
in your \code{.Rprofile}, then nothing will happen except both \code{.Rprofile}s will
be displayed in the Source pane.

To view the \code{.Rprofile} file in the \code{conr} package, run the following code in
the console:

\code{
  system.file(package = "conr", "preferences", ".Rprofile") |>
    file.edit()
}

The existing \code{.Rprofile} can be checked using
\code{\link[usethis]{edit_r_profile}(scope ="user")}.
}
\seealso{
\code{\link[=overwrite_preferences]{overwrite_preferences()}}
}
