% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.decline.fit.R
\name{pop.decline.fit}
\alias{pop.decline.fit}
\title{Fit Statistical Models to Population Reduction}
\usage{
pop.decline.fit(
  pop.size,
  years,
  models = "all",
  project.years = NULL,
  plot = TRUE
)
}
\arguments{
\item{pop.size}{a vector containing the (estimated) number of mature individuals of the species}

\item{years}{a vector containing the years for which the number of mature individuals was estimated}

\item{models}{a vector containing the names of the models to be fitted to species population data}

\item{project.years}{a vector containing the years for which the number of mature individuals should be predicted}

\item{plot}{logical. Should the fit of the best model be plotted against the population data?}
}
\description{
Fitting statistical models to the decline on the number of mature
 individuals across time "can be used to extrapolate population trends, so that 
 a reduction of three generations can be calculated" (IUCN 2019). This function
 provide a comparison of five different models and returns the model with best
 fit to data.
}
\details{
TO BE CHECKED By default, the function compare the fit of five statistical models to the population trends, 
 namely: linear, accelerating, exponential, logistic and piece-wise. Here, we use a quadratic model as
 a proxy AIC for time intervals > 30 and AICc for time intervals < 30.
 Note that model fits are performed in the proportional population estimates and on the period of interval,
 and not on raw population sizes and years of assessments. So, models should only be used to project 
 proportions.
 Fitting piece-wise models is very unstable (model fitting is sensitive to the start parameters) and may take a while to converge (use preferably for many years of assessment).
 We assumed that populations are stable or declining, so proportions are obtained in respect to the oldest population
 estimate.
 
 The model to be fitted should be based on the pattern of decline (which may be exponential, linear, accelerated, or a more complex pattern),
 which may be inferred from the type of threat. The assumed pattern of decline can make an important difference. Assessors should indicate the 
 basis on which they have decided the form of the decline function. The best information about the processes that contribute to changes in population 
 size should be used to decide what form of decline function to apply over the three-generation period. Specifically, if a model is fitted, the assumptions 
 of the model must be justified by characteristics of life history, habitat biology, pattern of exploitation or other threatening processes, etc. For example:
}
\examples{
pop = c(10000, 9100, 8200, 7500, 7000)
yrs = c(1970, 1975, 1980, 1985, 1990)
pop.decline.fit(pop.size = pop, years = yrs, models = c("linear"))
pop.decline.fit(pop.size = pop, years = yrs, models = c("linear", "exponential"))
pop.decline.fit(pop.size = pop, years = yrs, models = c("linear", "exponential"), project.years = c(1960, 2005))
pop.decline.fit(pop.size = pop, years = yrs, models = c("linear", "exponential"), project.years = c(1960, 2005), plot = FALSE)


### ERROR !
# AIC NA when using AICctab function
pop.decline.fit(pop.size = pop, years = yrs, models = c("general_logistic"))
# same problem 
pop.decline.fit(pop.size = pop, years = yrs, models = c("accelerating"))


}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Lima, R.A.F. & Dauby, G.
}
