% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.decline.R
\name{pop.decline}
\alias{pop.decline}
\title{Calculating Population Reduction}
\usage{
pop.decline(
  pop.size = NULL,
  years = NULL,
  taxa = NULL,
  models = "all",
  project.years = NULL,
  output = "all",
  by.taxon = FALSE,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{pop.size}{a vector, data frame or matrix containing the (estimated)
number of mature individuals of species (i.e. population size). If a data
frame or matrix, rows are the species and columns are the population sizes.}

\item{years}{a vector containing the years for which the population sizes is
available}

\item{taxa}{a vector containing the name of the species in \code{pop.size}}

\item{models}{a vector containing the names of the statistical models to be
fitted to species population data}

\item{project.years}{a vector containing the years for which the number of
mature individuals should be predicted using the best candidate statistical
model}

\item{output}{a character or vector containing the desired output from the
function. The options are: "predictions", "model.fit", "model.selection"
and "best.model". By default, the function returns only the precitions.}

\item{by.taxon}{logical. Should the output list be organized by the selected
output options (i.e. predictions, model.fit, model.selection and
best.model) for all taxa or should it cointain one taxon per taxon with
all selected outputs? Defaults to FALSE (list organized by outputs and not
taxa).}

\item{parallel}{a logical. Whether running should be performed in parallel.
FALSE by default.}

\item{NbeCores}{an integer. Register the number of cores for parallel
execution. Two by default.}

\item{show_progress}{logical. Whether progress informations should be
displayed. TRUE by default}

\item{...}{other parameters to be passed as arguments for functions
\code{pop.decline.fit} and \code{ICtab.mod.select}}
}
\value{
a named list
}
\description{
Based on the fit of statistical models to population data, the
function estimates the decline on the number of mature individuals across
time, expressed in percentage.
}
\details{
By default, the function compares the fit of six statistical models to the
population trends, namely: linear, quadratic, exponential, logistic,
generalized logistic and piece-wise. But users can use different combinations
of those models using the argument \code{models}, according to the specificity
of their study region or groups of organism. See \code{pop.decline.fit} for
more details and assumptions on how those models are fitted to population
data and how the candidate models are selected.
}
\examples{
## Creating vectors with the population data and time intervals 
#(adapted from the IUCN 2019 workbook for Criterion A, available 
#at: https://www.iucnredlist.org/resources/criterion-a)

pop = c(10000, 9050, 8250, 7500, 7200, 6950)
pop1 = c(10000, NA, 8200, NA, NA, 6000)
yrs = c(1970, 1975, 1980, 1985, 1990, 2000)
tax = c("species A", "species B")
pops = matrix(c(pop, pop1), nrow = length(tax), 
  dimnames = list(tax, yrs), byrow = TRUE)

## Fitting data with different models and settings
# only one species, all models (default)
pop.decline(pop, yrs)

# two species or more
pop.decline(pops)

# two species or more, less models 
pop.decline(pops, models = c("linear", "quadratic"))
pop.decline(pops, models = "exponential")

# two species or more, exponential models with projections
pop.decline(pops, models = "exponential", project.years = c(1960, 2050))
pop.decline(pops, models = "exponential", project.years = c(1973, 2005))

# two species or more, different outputs 
pop.decline(pops, models = "exponential", output = "model.fit")
pop.decline(pops, models = c("linear", "quadratic", "exponential"), 
 output = "model.selection")

## Another examples 
# Few observations (warning or no model fit below 3 observations) 
pop.decline(pop.size = c(10000, 8200, 6000), years = c(1970, 1985, 2000), 
 models = "all", project.years = 2030)

\dontrun{
# Not enough observations (error)
pop.decline(pop.size = c(10000, 6000), years = c(1970, 2000))
}


}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and
Criteria. Version 14. Standards and Petitions Committee. Downloadable from:
http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\seealso{
\link[ConR]{pop.decline.fit}
}
\author{
Renato A. Ferreira de Lima
}
