% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_regrsn.R
\name{print_regrsn}
\alias{print_regrsn}
\title{Print Regression Summary Table}
\usage{
print_regrsn(models = list(), accuracy = 1, omit = c("Num.Obs"))
}
\arguments{
\item{models}{Regression model(s).}

\item{accuracy}{Argument in scales::label_number(accuracy = 1).}

\item{omit}{Argument in modelsummary::modelsummary(gof_omit = omit).}
}
\value{
flextable::flextable.
}
\description{
Use modelsummary and flextable to produce a regression table.
}
