% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Produce Simple Quantile-Quantile Plots}
\usage{
qq_plot(model)
}
\arguments{
\item{model}{}
}
\value{
A ggplot.
}
\description{
Produce Simple Quantile-Quantile Plots
}
\examples{
my_model <- lm(y ~ x, data)
qq_plot(my_model)
}
