% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_sensibly.R
\name{round_sensibly}
\alias{round_sensibly}
\title{Round numbers sensibly}
\usage{
round_sensibly(x, digits = 0)
}
\arguments{
\item{x}{The number you intend to round.}

\item{digits}{The number of digits after the decimal, defaults to zero.}
}
\value{
A double.
}
\description{
base::round() has a flaw in that it rounds 5 to the nearest even
number - i.e. 1.5 is rounded to 2 but 2.5 is also rounded to 2. This
function was originally stolen from
\url{https://stackoverflow.com/questions/12688717/round-up-from-5} and I
subsequently found \code{janitor::round_half_up}. Added exception to the
\code{\link[lubridate]{timespan}} class, which is numeric but can't be
rounded.
}
\examples{
base::round(2.5)
conr::round_sensibly(2.5)
conr::round_sensibly(2.115, 2)

\dontshow{if (requireNamespace("lubridate")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(lubridate)
lubridate::interval(
  lubridate::ymd("2021-01-01"),
  lubridate::ymd("2021-04-01")
) |>
  round_sensibly()
\dontshow{\}) # examplesIf}
}
