% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scripts.R
\name{run_scripts}
\alias{run_scripts}
\title{Run Scripts}
\usage{
run_scripts(directory = here::here("code"), last = Inf)
}
\arguments{
\item{directory}{Character. What directory contains the scripts to run? The
default is in the \code{code} subdirectory of the open project.}

\item{last}{Numeric. What is the last script to run. Default is \code{Inf}. If
later scripts are a work in progress, you don't want to run them.}
}
\value{
Whatever your scripts return after being parsed to \code{source}.
}
\description{
Convenience function to run all (or some) scripts in a
subdirectory. The function currently assumes all scripts to be run are
named similar to \verb{01_clean.R}, with numbers somewhere in the name to
indicate ordering.
}
