% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop.radius.R
\name{subpop.radius}
\alias{subpop.radius}
\title{Estimate Taxon Dispersal Radius}
\usage{
subpop.radius(
  XY,
  factor.div = 10,
  quant.max = 1,
  mode = "spheroid",
  proj_type = "cea"
)
}
\arguments{
\item{XY}{a data.frame. containing the geographical coordinates for each taxon.}

\item{factor.div}{numeric. denominator value used to obtain the fraction of
the maximum distance. Default to 10.}

\item{quant.max}{numeric. The upper-quantile of the inter-point distance to
be considered as a threshold of maximum distance. Can vary between 0 and 1.
Default to 1.}

\item{mode}{a character. Type of coordinate projection: "spheroid" or
"planar". Defualts to "spheroid".}

\item{proj_type}{string or numeric}
}
\value{
The estimated radius in kilometres for each taxon.
}
\description{
Calculates the fraction of the maximum distance between
species occurrences, which can be used as a proxy of species dispersal
ability. This fraction can be used as the radius needed to apply the
circular buffer method (Rivers et al 2010) to obtain subpopulations for
conservation assessments of extinction risk.
}
\details{
\code{XY} as a \link[base:data.frame]{data.frame} should have the following structure:

\strong{It is mandatory to respect field positions, but field names do not matter}

\enumerate{
\item The first column is contains numeric value i.e. latitude in decimal degrees
\item The second column is contains numeric value i.e. longitude in decimal degrees
\item The third column is contains character value i.e. the names of the species
}
}
\examples{

mydf <- data.frame(ddlat = c(-44.6,-46.2,-45.4,-42.2,-43.7,-45.0,-28.0),
                   ddlon = c(-42.2,-42.6,-45.3,-42.5,-42.3,-39.0,-17.2),
                   tax = rep("a", 7),
                   stringsAsFactors = FALSE)
                   
subpop.radius(mydf)
subpop.radius(mydf, quant.max = 0.95)
subpop.radius(mydf, factor.div = 15)

}
\references{
Rivers MC, Bachman SP, Meagher TR, Lughadha EN, Brummitt NA
(2010) Subpopulations, locations and fragmentation: applying IUCN red list
criteria to herbarium specimen data. Biodiversity and Conservation 19:
2071-2085. doi: 10.1007/s10531-010-9826-9
}
\author{
Renato A. Ferreira de Lima & Gilles Dauby
}
