% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_parallel.R
\name{activate_parallel}
\alias{activate_parallel}
\title{Internal function}
\usage{
activate_parallel(parallel = FALSE, NbeCores = 2)
}
\arguments{
\item{parallel}{a logical. Whether running should be performed in parallel.
FALSE by default.}

\item{NbeCores}{an integer. Register the number of cores for parallel
execution. Two by default.}
}
\description{
Activate parallel processing
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
\keyword{internal}
