% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_df.R
\name{adorn_df}
\alias{adorn_df}
\title{Adorn a Dataframe with Percent and '000s Separator}
\usage{
adorn_df(df, perc_digits = 1, digits = 1, ...)
}
\arguments{
\item{df}{A dataframe (tibble expected).}

\item{perc_digits}{A number, like in format(accuracy = _), for percentage
columns.}

\item{digits}{A number like perc_digits, but for numeric columns.}

\item{...}{}
}
\value{
A dataframe.
}
\description{
Adorn a Dataframe with Percent and '000s Separator
}
