% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOH.estimation.R
\name{AOH.estimation}
\alias{AOH.estimation}
\title{Terrestrial Area of Habitat}
\usage{
AOH.estimation(
  XY,
  show_progress = TRUE,
  proj_type = "cea",
  mode = "planar",
  hab.map = NULL,
  hab.map.type = NULL,
  hab.class = NULL,
  years = NULL,
  country_map = NULL,
  exclude.area = TRUE,
  parallel = FALSE,
  NbeCores = 2,
  simplifiy_poly = TRUE,
  buffer = NULL,
  ...
)
}
\arguments{
\item{XY}{\code{"dataframe"} see Details}

\item{show_progress}{logical. whether a bar showing progress in computation
should be shown. Default to TRUE.}

\item{proj_type}{character string, numeric or object of CRS class. Default to
"cea".}

\item{mode}{character string either 'spheroid' or 'planar'. By default
'spheroid'}

\item{hab.map}{raster, raster layer/stack or spatial polygons containing the
habitat spatial information}

\item{hab.map.type}{logical, vector of same lenght of hab.map, if TRUE means
hab.map is suitable for species, if FALSE unsuitable}

\item{hab.class}{classes of values in \code{hab.map} to be considered as
'habitat'.}

\item{years}{numeric. Time interval between the first and last \code{hab.map}
if more than one raster is provided (e.g. if \code{hab.map} is a RasterStack
object).}

\item{country_map}{a \code{SpatialPolygonsDataFrame} or}

\item{exclude.area}{a logical, if TRUE, areas outside of \code{country_map} are
cropped of \code{SpatialPolygons} used for calculating EOO. By default is FALSE}

\item{parallel}{logical. Should computing run in parallel? Default to FALSE.}

\item{NbeCores}{integer. The number of cores for parallel execution. Default
to 2.}

\item{simplifiy_poly}{logical whether the resulting polygon should be
simplified using ms_simplify function of rmapshaper package}

\item{buffer}{logical}

\item{...}{additional arguments passed to \code{EOO.computing}

\code{SpatialPolygons} showing for example countries or continent borders.
This shapefile will be used for cropping the \code{SpatialPolygons}l if
exclude.area is TRUE}
}
\value{
If \code{hab.map} is a *Raster object with categorical variables, the
function returns a data frame containing the taxon name (tax), the
percentage of habitat within the taxon EOO (prop.EOO) and the approximate
area that this percentage represents. If habitat change is also estimated
the function returns additional columns, namely:
\itemize{
\item the total percentage loss and recover of habitat within the EOO (loss,
recover);
\item the percentage loss and recover relative to the amount of habitat and
non-habitat within the EOO at the beginning of the time interval (rel.loss,
rel.recover);
\item the annual rate of habitat loss and recover (rate.loss, rate.recover,
i.e. rel.loss and rel.recover divided by \code{years});
\item A habitat quality index which is obtained by averaging values of quality
to pixels according to their transition between the first and last year of
the times series, namely habitat to habitat = 2, non-habitat to habitat =
1, non-habitat to non-habitat = 0, and habitat to non-habitat = -1
(hab.quality, hab.quality.lo, hab.quality.hi). The index thus vary from -1
(i.e., all pixels were habitat in time one and all was lost in time two) until 2
(i.e., all pixels were habitat in time one and they remained so in time 2).
}

If \code{hab.map} is a *Raster object with continuous variables, the
function returns a data frame containing the taxon name (tax), and, by
default, the summary statistics of all pixels within the taxon EOO. If
argument \code{output_raster} is "prop.table" the function return the proportion
of each pixel value within the taxon EOO and if it is "area.table" it returns
the approximate area of each pixel value within the taxon EOO.

If \code{hab.map} is a 'sp' or 'sf' object with multiple polygons, the
function returns a data frame containing the taxon name (tax), the number
of unique polygons (numb.polys), the percentage of area (prop.EOO) and the
total area of the polygons within the taxon EOO.

If \code{export_shp} is TRUE, the function also returns the transition rasters
used to calculated 'hab.quality' or the cropped polygons in the case of
polygons. But note that for rasters, this include additional steps that can
be time and memory consuming.
}
\description{
The function compute the amount of 'habitat' within the EOO of
each species, which may represent habitat itself or any other proxy, value or
category the user may find relevant (e.g. area covered by protected areas,
altitude, etc...). If two or more time intervals are provided, the function
also returns the habitat change between intervals. Currently, the type of
habitat maps objects currently accepted are sp, sf and Raster*. In
addition, habitat change is only computed for rasters (e.g. RasterStack
object).
}
\details{
If multiple rasters are provided, the last raster is taken as the
most recent one to calculate current habitat proportion and area.
Similarly, the first and last rasters are used to calculate habitat change.
In this case the user should provide the interval in years between the two
rasters. If this interval is not provided the function assumes the number
of rasters provided minus one as the 'default' time interval.

For faster processing of maps, the function automatically crops the EOO
polygons to the extent of the \code{hab.map} provided. Then, for each taxon,
the \code{hab.map} is cropped to the extent of the taxon EOO polygon. If the
taxon EOO and the raster provided don't intersect the functions returns NA.

If the size of the EOO is too small in respect to the raster resolution
(small enough to contain an entire pixel) or if the EOO polygon is
elongated (perimeter/area relationship > 2), the function returns the
estimates for all pixels touching the species EOO boundaries and not only
the pixels within the EOO. However, if the EOO polygon is small enough to
fit entirely within a pixel, metrics will refer to this single pixel. So,
make sure that the resolution of the rasters provided make sense to the
range of the taxa being assessed.
}
