% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO.computing.R
\name{AOO.computing}
\alias{AOO.computing}
\title{Area of occupancy}
\usage{
AOO.computing(
  XY,
  cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  export_shp = FALSE,
  proj_type = "cea"
)
}
\arguments{
\item{XY}{\link[base:data.frame]{data.frame} see Details.}

\item{cell_size_AOO}{numeric, by default is 2. Value indicating the grid size
in kilometres used for estimating Area of Occupancy.}

\item{nbe.rep.rast.AOO}{numeric, by default is 0. Indicate the number of
raster with random starting position used for estimating the AOO. If 0 but
some translation of the raster are still done.}

\item{parallel}{logical, by default is FALSE. Whether running in parallel.}

\item{NbeCores}{integer, by default is 2. Register the number of cores for
parallel execution. Only used if parallel is TRUE.}

\item{show_progress}{logical, by default is TRUE. Whether a progress bar
during computation is shown.}

\item{export_shp}{logical, by default is FALSE. Whether a shapefile of
occupied cells should be exported.}

\item{proj_type}{character or numeric, by default is "cea", see Details.}
}
\value{
If \code{export_shp} if FALSE (the default) a numeric vector of AOO estimates for each taxa
If \code{export_shp} if TRUE a list with two elements
\enumerate{
\item a vector of AOO estimates for each taxa
\item a list of simple feature for each taxa
}
}
\description{
Compute areas of occupancy (AOO) for multiple taxa in square kilometers
}
\section{Input data}{
\strong{XY} as a \link[base:data.frame]{data.frame} should have the following structure:

\strong{It is mandatory to respect field positions, but column names do not matter}

\enumerate{
\item The first column is contains numeric value i.e. latitude in decimal degrees
\item The second column is contains numeric value i.e. longitude in decimal degrees
\item The third column is contains character value i.e. the names of the species
}

See Examples.
}

\section{Iteration to get the minimal AOO}{
The argument \code{nbe.rep.rast.AOO} should ideally be higher than 10 for
increasing the chance to get the minimal number of occupied cell. However,
increasing \code{nbe.rep.rast.AOO} also increases the computing time. Note that if
\code{nbe.rep.rast.AOO = 0}, several translations of the grid overlaying
occurrences are still conducted
}

\section{proj_type}{
See \code{proj_type}
}

\examples{
data(dataset.ex)

AOO <- AOO.computing(dataset.ex)


# This would estimate AOO for all taxa by overlaying randomly a 
# grid 10 times. For each taxa, the minimum value is kept

AOO <- AOO.computing(dataset.ex, nbe.rep.rast.AO = 10)


}
\references{
Gaston & Fuller 2009 The sizes of species'geographic ranges,
Journal of Applied Ecology, 49 1-9
}
\author{
Gilles Dauby \email{gilles.dauby@ird.fr}
}
