% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOO.decline.R
\name{AOO.decline}
\alias{AOO.decline}
\title{Area of occupancy decline}
\usage{
AOO.decline(
  XY,
  hab.map,
  cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  proj_type = "cea",
  hab.class = NULL,
  hab.map.type = NULL,
  all_individual_layers = FALSE
)
}
\arguments{
\item{XY}{\link[base:data.frame]{data.frame} see Details.}

\item{hab.map}{SpatRaster or sf polygons containing the
habitat spatial information}

\item{cell_size_AOO}{numeric, by default is 2. Value indicating the grid size
in kilometres used for estimating Area of Occupancy.}

\item{nbe.rep.rast.AOO}{numeric, by default is 0. Indicate the number of
raster with random starting position used for estimating the AOO. If 0 but
some translation of the raster are still done.}

\item{parallel}{logical, by default is FALSE. Whether running in parallel.}

\item{NbeCores}{integer, by default is 2. Register the number of cores for
parallel execution. Only used if parallel is TRUE.}

\item{show_progress}{logical, by default is TRUE. Whether a progress bar
during computation is shown.}

\item{proj_type}{character or numeric, by default is "cea", see Details.}

\item{hab.class}{classes of values in \code{hab.map} to be considered as suitable}

\item{hab.map.type}{logical, vector of same length of \code{hab.map},
\itemize{
\item \code{TRUE} means the habitat of \code{hab.map} is suitable
\item \code{FALSE} means the habitat of \code{hab.map} is unsuitable
}}

\item{all_individual_layers}{logical
\itemize{
\item \code{TRUE} compute AOO decline for each individual \code{hab.map}
\item \code{FALSE} (the default) compute AOO decline for all layers together
}}
}
\value{
\enumerate{
\item AOOs a \code{dataframe} of AOO estimates for each taxa and for each layer/spatial polygons if \code{all_individual_layers} is TRUE
\item AOO_decline a \code{dataframe} of AOO.decline in percentages
\item categories based on the sub-criteria of IUCN criterion A
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Estimate areas of occupancy (AOO) decline for multiple taxa in square kilometres and percentage
}
\details{
\strong{Input} as a \link[base:data.frame]{data.frame} should have the following structure:

\strong{It is mandatory to respect field positions, but field names do not matter}\tabular{lll}{
   latitude \tab longitude \tab species \cr
   numeric \tab numeric \tab character \cr
}


The argument \code{nbe.rep.rast.AOO} should ideally be higher than 20 for increasing
the chance to get the minimal number of occupied cell. Increasing \code{nbe.rep.rast.AOO} however
also increase the computing time. So this is a trade-off that depends on the importance to
get the minimal AOO and the size of the dataset.
}
\examples{

m <- matrix(1:25, nrow=5, ncol=5)
rm <- terra::rast(m)
terra::values(rm) <- sample(c("forest", "cities", "roads"), 25, replace = TRUE)
cls <- data.frame(id=1:3, cover=c("forest", "cities", "roads"))
levels(rm) <- cls
terra::crs(rm) <- "epsg:4326"


test_data <- dummy_dist(n = 5, xmin = 0, xmax = 5, ymin = 0, ymax = 5)


res <- AOO.decline(
XY = test_data,
hab.map = rm,
hab.class = c("forest"),
all_individual_layers = TRUE
)


res <- AOO.decline(
XY = test_data,
hab.map = rm,
hab.class = c("cities", "roads"),
all_individual_layers = TRUE, 
hab.map.type = FALSE ### this means the provided hab.map is unsuitable
)


}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
