% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_criterion_a.R
\name{cat_criterion_a}
\alias{cat_criterion_a}
\title{Categorize taxa according to IUCN criterion A}
\usage{
cat_criterion_a(
  A1_val = NULL,
  A2_val = NULL,
  A3_val = NULL,
  A4_val = NULL,
  A1.threshold = c(50, 70, 90),
  A234.threshold = c(30, 50, 80),
  all.cats = TRUE
)
}
\arguments{
\item{A1_val}{a numeric vector of the estimates of population decline based on IUCN sub-criterion A1}

\item{A2_val}{a numeric vector of the estimates of population decline based on IUCN sub-criterion A2}

\item{A3_val}{a numeric vector of the estimates of population decline based on IUCN sub-criterion A3}

\item{A4_val}{a numeric vector of the estimates of population decline based on IUCN sub-criterion A4}

\item{A1.threshold}{a numeric vector of threshold values for sub-criterion A1, default is c(50, 70, 90) following IUCN guidelines}

\item{A234.threshold}{a numeric vector of A2, A3 and A4 threshold values for sub-criteria A2, A3 and A4, default is c(30, 50, 80) following IUCN guidelines}

\item{all.cats}{a logical value indicating whether to return all categories or only the consensus, default is TRUE}
}
\value{
A list of two or three elements:
\itemize{
\item \code{ranks_A}: a character vector with the consensus category of each taxon based on subcriteria A1-A4
\item \code{cats_code}: a character vector with the final IUCN category for each taxon
\item \code{all_cats}: a data frame with the categorization of each taxon based on all sub criteria, only returned if \code{all.cats=TRUE}
}
}
\description{
Provide the consensus IUCN category based on the sub-criteria of
IUCN criterion A (A1, A2, A3 and A4) and the thresholds recommended by IUCN.
}
\details{
By default, the function provides the consensus category, following the recommendations of IUCN (2019)
that states "Only the criteria for the highest category of threat that the taxon qualifies for should be listed".
Therefore, the consensus category is the highest category of threat among the sub-criteria evaluated.

The function assumes that the order of the values in A1_val, A2_val, A3_val and A4_val are
from the same taxa (i.e. first element from A1_val until A4_val is always the same species i).
Therefore, the order of the estimates of population decline for each sub-criterion \emph{must} be the same.
}
\examples{
A1_val <- c(40, 90, 75, 10)
A2_val <- c(90, 95, 30, 10)
A3_val <- c(10, 30, 45, 15)
A4_val <- c(32, 18, 80, 20)

## All four criteria
cat_criterion_a(
  A1_val = A1_val,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val
)

## All four criteria, no categories for all criteria, just the consesus
cat_criterion_a(
  A1_val = A1_val,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val,
  all.cats = FALSE
)

## One or more criteria not evaluated
cat_criterion_a(
  A1_val = NULL,
  A2_val = A2_val,
  A3_val = A3_val,
  A4_val = A4_val
)

cat_criterion_a(
  A1_val = A1_val,
  A2_val = NULL,
  A3_val = NULL,
  A4_val = NULL
)

## One or more criteria not evaluated, one species
cat_criterion_a(
  A1_val = NULL,
  A2_val = 0.4,
  A3_val = NULL,
  A4_val = NULL
)

}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and
Criteria. Version 14. Standards and Petitions Committee. Downloadable from:
http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Gilles Dauby & Renato A. Ferreira de Lima
}
