% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_criterion_b.R
\name{cat_criterion_b}
\alias{cat_criterion_b}
\title{Categorize taxa according to IUCN criterion B}
\usage{
cat_criterion_b(
  EOO = NULL,
  AOO = NULL,
  locations = NULL,
  sever.frag = NULL,
  protected = NULL,
  decline = NULL,
  ext.fluct = NULL,
  EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10),
  Loc.threshold = c(10, 5, 1),
  protected.threshold = 100,
  fluct.threshold = 10,
  all.cats = TRUE
)
}
\arguments{
\item{EOO}{numeric vector with species extent of occurrence - EOO (i.e.
sub-criterion B1)}

\item{AOO}{numeric vector with species area of occupancy - AOO (i.e.
sub-criterion B2)}

\item{locations}{numeric vector with the number of locations where the
species occur (i.e. condition 'a' of criterion B)}

\item{sever.frag}{logical. Is the population severely fragmented? If TRUE 
the condition 'a' of criterion B will be met.}

\item{protected}{numeric vector providing estimated percentage of species
distribution in protected areas}

\item{decline}{string vector providing the status of the species continuing
decline in EOO, AOO, habitat, locations or subpopulations or population
size (i.e. condition 'b'). If different of 'Decreasing', the condition 'b'
of criterion B will not be met.}

\item{ext.fluct}{numeric vector with the mean order of magnitude of the
differences between population minima and maxima (Currently not
implemented).}

\item{EOO.threshold}{numeric vector with the EOO thresholds to convert
estimates into threat categories. Default is the thresholds recommended by
IUCN.}

\item{AOO.threshold}{numeric vector with the AOO thresholds to convert
estimates into threat categories. Default is the thresholds recommended by
IUCN.}

\item{Loc.threshold}{numeric vector with the thresholds of number of
locations (condition 'a'). Default is the thresholds recommended by IUCN.}

\item{protected.threshold}{numeric, one value indicating the threshold for
protected value above which a taxa would not be threatened whatever the
others parameters, by default is 100}

\item{fluct.threshold}{numeric. Threshold of the order of magnitude of the
differences between population minima and maxima to classify extreme
fluctuations. Default to 10 as recommended by IUCN.}

\item{all.cats}{logical. Should the categories from all criteria be returned
and not just the consensus categories? Default to TRUE.}
}
\value{
list
}
\description{
Provide IUCN threat categories based on B sub-criteria,
  conditions and thresholds.
}
\details{
The function categorizes taxa following criterion B and categories of the IUCN.
}
\examples{

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(9, 12)
cat_criterion_b(EOO = EOO,AOO = AOO, locations = locations)

EOO <- c(34000, 80)
AOO <- c(300, 25)
locations <- c(1, 1)
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations)

EOO <- c(50, 5000)
AOO <- c(5, 25)
locations <- c(1, 10)
protected <- c(80, 50)
decline <- c("Stable", "Decreasing")
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations, protected = protected, decline = decline)

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(9, 12)
sever.frag <- c(FALSE, TRUE)
protected <- c(100, 80)
decline <- c("Decreasing", "Decreasing")
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations, sever.frag = sever.frag, protected = protected, decline = decline)

EOO <- c(34000, 5000)
AOO <- c(300, 25)
locations <- c(15, 12)
sever.frag <- c(TRUE, TRUE)
protected <- c(80, 50)
decline <- c("Decreasing", "Decreasing")
cat_criterion_b(EOO = EOO, AOO = AOO, locations = locations, sever.frag = sever.frag, protected = protected, decline = decline)

}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and
  Criteria. Version 14. Standards and Petitions Committee. Downloadable from:
  http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Dauby, G. & Lima, R.A.F.
}
