\name{ConR-package}
\alias{ConR-package}
\alias{ConR}
\docType{package}
\title{
\packageTitle{ConR}
}
\description{
 R functions for calculating statistics used for a preliminary \cr
	conservation assessment following Criterion B of the IUCN (Extent of Occurrence, \cr
	Area of Occupancy, number of locations, number of sub-populations)
}
\details{
\packageIndices{ConR}
}
\author{
\packageAuthor{ConR}

Maintainer: \packageMaintainer{ConR}
}


\keyword{ package }


