% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convex.Hull.Poly.R
\name{Convex.Hull.Poly}
\alias{Convex.Hull.Poly}
\title{Internal function}
\usage{
Convex.Hull.Poly(XY, mode = "spheroid", proj_type = NULL)
}
\arguments{
\item{XY}{data.frame}

\item{mode}{character string either 'spheroid' or 'planar'. By default 'spheroid'}

\item{proj_type}{crs}
}
\description{
Build convex hull polygon
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
