% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.check.R
\name{coord.check}
\alias{coord.check}
\title{Internal function}
\usage{
coord.check(XY, listing = TRUE, proj_type = NULL, listing_by_valid = FALSE)
}
\arguments{
\item{XY}{data.frame, of at least two columns (coordinates), third is taxa}

\item{listing}{logical, whether the dataset should be splitted in a list by taxa}

\item{proj_type}{...}

\item{listing_by_valid}{...}
}
\description{
Coordinates check
}
