% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.check.R
\name{coord.check}
\alias{coord.check}
\title{Internal function}
\usage{
coord.check(
  XY,
  listing = TRUE,
  proj_type = NULL,
  listing_by_valid = FALSE,
  cell_size = NULL,
  check_eoo = TRUE
)
}
\arguments{
\item{XY}{data.frame, of at least two columns (coordinates), third is taxa}

\item{listing}{logical, whether the dataset should be splitted in a list by taxa}

\item{proj_type}{crs}

\item{listing_by_valid}{logical}

\item{check_eoo}{logical}
}
\value{
a list with verified dataset and added columns
}
\description{
Coordinates check
}
\examples{

dataset <- coord.check(XY = dummy_dist())
dataset$list_data
dataset$issue_nrow ## this indicates which species has less that 3 unique pairs of coordinates

}
