% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_A.R
\name{criterion_A}
\alias{criterion_A}
\title{Assess IUCN Criterion A}
\usage{
criterion_A(
  x,
  years = NULL,
  assess.year = NULL,
  project.years = NULL,
  generation.time = NULL,
  models = c("linear", "accelerating", "exponential", "logistic", "general_logistic"),
  A1.threshold = c(50, 70, 90),
  A234.threshold = c(30, 50, 80)
)
}
\arguments{
\item{x}{a vector or a data frame containing the (estimated) number of mature individuals of the species}

\item{years}{a vector containing the years for which the number of mature individuals was estimated}

\item{assess.year}{}

\item{project.years}{a vector containing the years for which the number of mature individuals should be predicted}

\item{generation.time}{}

\item{models}{a vector containing the names of the models to be fitted to species population data}

\item{A1.threshold}{}

\item{A234.threshold}{}
}
\value{
TO BE COMPLETED
}
\description{
Preliminary assessment of species conservation status following
 IUCN Criterion A, which is based on population size reductions (Criteria A1, 
 A2, A3 and A4)
}
\details{
TO BE COMPLETED
}
\examples{
data(example_criterionA)
criterion_A(example_criterionA, years = NULL, assess.year = 2000, project.years = seq(2002, 2030, by = 2), generation.time = 10)

}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and Criteria. Version 14. Standards and Petitions Committee. Downloadable from: http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Lima, R.A.F. & Dauby, G.
}
