% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_B.R
\name{criterion_B}
\alias{criterion_B}
\title{Assess IUCN Criterion B}
\usage{
criterion_B(
  x,
  protec.areas = NULL,
  NamesSp = "species1",
  file_name = NULL,
  EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10),
  Loc.threshold = c(10, 5, 1)
)
}
\arguments{
\item{x}{the character string.}
}
\value{

}
\description{
Preliminary assessment of species conservation status following
 IUCN Criterion B, which is based on species geographic distribution (i.e. extent
 of occurrence - EOO, and area of occupancy, AOO)
}
\details{
The function ...
}
\examples{


}
\references{

}
\author{
Dauby, G. & Lima, R.A.F.
}
