% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_B.R
\name{criterion_B}
\alias{criterion_B}
\title{Assess IUCN Criterion B}
\usage{
criterion_B(
  x = NULL,
  AOO = NULL,
  EOO = NULL,
  locations = NULL,
  severe.frag = NULL,
  subpops = NULL,
  EOO.threshold = c(20000, 5000, 100),
  AOO.threshold = c(2000, 500, 10),
  Loc.threshold = c(10, 5, 1),
  comp_subpop = TRUE,
  comp_severe.frag = FALSE,
  resol_sub_pop = 5,
  cell_size_locations = 10,
  method_locations = "fixed_grid",
  method_polygons = "no_more_than_one",
  rel_cell_size = 0.05,
  threat_list = NULL,
  names_threat = NULL,
  threat_weight = NULL,
  id_shape = "id_orig",
  country_map = NULL,
  method.range = "convex.hull",
  alpha = 1,
  buff.alpha = 0.1,
  exclude.area = FALSE,
  cell_size_AOO = 2,
  nbe.rep.rast.AOO = 0,
  threshold_severe = 50,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2,
  proj_type = "cea",
  mode = "spheroid",
  DrawMap = FALSE,
  add.legend = TRUE
)
}
\arguments{
\item{x}{a \code{dataframe} or an object of class \code{spgeoIN} see
\url{https://github.com/azizka/speciesgeocodeR}. See Details}

\item{EOO.threshold}{numeric vector indicating the thresholds used to categorize EOO in IUCN categories}

\item{AOO.threshold}{numeric vector indicating the thresholds used to categorize AOO in IUCN categories}

\item{Loc.threshold}{numeric vector indicating the thresholds used to categorize the number of locations in IUCN categories}

\item{comp_subpop}{logical if the number of sub-populations should be computed. By default is TRUE}

\item{resol_sub_pop}{a value defining the radius of the circles around each
occurrence (in kilometres) or data frame vector containing a column 'tax'
with the taxa names and a column 'radius' with the species-specific radius
(in kilometre as well). Typically, this data frame is the output of
\code{ConR} function \code{subpop.radius}.}

\item{cell_size_locations}{numeric, value indicating the grid size in kilometres used for estimating the number of location. By default, equal to 10}

\item{method_locations}{string, indicating the method used for estimating the number of locations. See Details
\itemize{
\item \code{"fixed_grid"} (by default)
\item \code{"sliding_scale"}
}}

\item{method_polygons}{string. Used if \code{threat_list} is provided. See Details.
\itemize{
\item \code{"no_more_than_one"} (the default): each single POLYGON will be considered as a single location
\item \code{"grid"}: a grid of \code{cell_size_locations} size will be used to estimate the number of location within polygons
}}

\item{rel_cell_size}{numeric, if \code{method_locations="sliding_scale"}, \code{cell_size_locations} is ignored
and the resolution is given by the maximum distance separating two occurrences multiplied by \code{rel_cell_size}. By default, it is 0.05}

\item{threat_list}{list or sfc objects POLYGON or MULTIPOLYGON documenting threats. If provided, this will be taken into account for calculating number of location (see Details and \code{method_polygons}). By default, no shapefile is provided}

\item{names_threat}{character vector, indicating names of threats, optional}

\item{threat_weight}{numeric vector, indicating weight given to each threat}

\item{id_shape}{string}

\item{country_map}{a \code{SpatialPolygonsDataFrame} or
\code{SpatialPolygons} showing for example countries or continent borders.
This shapefile will be used for cropping the \code{SpatialPolygons}l if
exclude.area is TRUE}

\item{method.range}{a character string, "convex.hull" or "alpha.hull". By
default is "convex.hull"}

\item{alpha}{a numeric, if \code{method.range} is "alpha.hull", value of
alpha of the alpha hull, see \code{\link[alphahull:ahull]{alphahull::ahull()}}. By default is 1}

\item{buff.alpha}{a numeric, if \code{method.range} is "alpha.hull", define
the buffer in decimal degree added to alpha hull. By default is 0.1}

\item{exclude.area}{a logical, if TRUE, areas outside of \code{country_map}
are cropped of \code{SpatialPolygons} used for calculating EOO. By default
is FALSE}

\item{cell_size_AOO}{numeric, by default is 2. Value indicating the grid size
in kilometres used for estimating Area of Occupancy.}

\item{nbe.rep.rast.AOO}{numeric, by default is 0. Indicate the number of
raster with random starting position used for estimating the AOO. If 0 but
some translation of the raster are still done.}

\item{threshold_severe}{numeric of one value indicates the threshold
to identify severe fragmentation. By default is 50}

\item{parallel}{logical, whether running in parallel. By default, it is FALSE}

\item{show_progress}{logical, whether a bar showing progress in computation should be shown. By default, it is TRUE}

\item{NbeCores}{integer, register the number of cores for parallel execution. By default, it is 2}

\item{proj_type}{string or numeric}

\item{mode}{character string either 'spheroid' or 'planar'. By default 'spheroid'}

\item{DrawMap}{logical, by default is FALSE, if TRUE, a png map is created for each species
in a directory of the working environment}

\item{add.legend}{logical, whether legend should be added to map}
}
\value{
A data frame containing, for each of taxon, (EOO, AOO, n.locs, n.subpops?),
the IUCN categories associated with the sub-criteria and the consensus category
for criterion B.
}
\description{
Preliminary assessment of species conservation status following
IUCN Criterion B, which is based on species geographic distribution (i.e. extent
of occurrence - EOO, and area of occupancy, AOO)
}
\details{
The function ...
}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and
Criteria. Version 14. Standards and Petitions Committee. Downloadable from:
http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Gilles Dauby & Renato A. Ferreira de Lima
}
