% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_D.R
\name{criterion_D}
\alias{criterion_D}
\title{Assess IUCN Criterion D}
\usage{
criterion_D(
  pop.size = NULL,
  name_sp = NULL,
  AOO = NULL,
  n.Locs = NULL,
  prop.mature = NULL,
  subcriteria = c("D", "D2"),
  D.threshold = c(1000, 250, 50),
  AOO.threshold = NULL,
  Loc.threshold = NULL,
  all.cats = TRUE
)
}
\arguments{
\item{pop.size}{a value (one species) or a vector (multiple species/
subpopulations) containing the population sizes (e.g. number of mature
individuals).}

\item{name_sp}{a vector containing the name of the taxa to be assessed.
Default to "Species 1", "Species 2", ..., "Species n", where n is the
number of taxa.}

\item{AOO}{a vector containing the Area of Occupancy of each taxon (in km2).}

\item{n.Locs}{a vector containing the number of locations for each taxon.}

\item{prop.mature}{a value or vector of the proportion of mature individuals in the
total population (IUCN 2019). Default to 1.}

\item{subcriteria}{a vector containing the sub-criteria that should be
included in the assessment (i.e. D and/or D2).}

\item{D.threshold}{numeric vector with the criterion D thresholds of very
small population sizes. Default values are the thresholds recommended by
the IUCN.}

\item{AOO.threshold}{numeric vector containing the threshold of the Area of
Occupancy (AOO) in km2. Default to NULL.}

\item{Loc.threshold}{numeric vector containing the threshold of the Number of
Locations. Default to NULL.}

\item{all.cats}{logical. Should the categories from all criteria be returned
and not just the consensus categories?}
}
\value{
A data frame containing the name of each taxon, the population sizes,
the AOO and Number of Locations, the IUCN categories associated with each
sub-criterion and the consensus category for criterion D.
}
\description{
Preliminary assessment of species conservation status following
IUCN Criterion D, which is based on very small or restricted populations
(Subcriteria D1 and D2).
}
\details{
This is a simple and fast function to assess IUCN criterion D. The
assessment based solely on population sizes is done automatically, if a
vector of population sizes is provided. Caution should be taken while
assessing sub-criterion D2. IUCN (2019, p.71) emphasizes that there is no
strict thresholds for D2 and that this sub-criterion should only be
assessed if the "population is prone to the effects of human activities or
stochastic events in an uncertain future, and is thus capable of becoming
Critically Endangered or even Extinct in a very short time period. (...).
So, simply meeting the suggested (or any other) threshold for AOO or number
of locations is not sufficient". Therefore, \code{criterion_D} assumes no default
thresholds for D2.

The argument \code{prop.mature} can be used if the population data provided are not
already the number of mature individuals (i.e. population size sensu IUCN, 2019).
By default, the proportion of mature individuals in the total population proportion
is taken as 1, but the user can provide one proportion for all species or species-
specific proportions.

Currenlty, the function does not supports data separated by subpopulation.
}
\examples{

## All sub-criteria
criterion_D(pop.size = c(25, 50, 200, 250, 500, 1000, 1500), 
            AOO = c(1,2,5,10,15,20,25), 
            n.Locs = c(1,1,2,3,4,5,6),
            subcriteria = c("D", "D2"),
            AOO.threshold = 20,
            Loc.threshold = 5)
            
## Only sub-criterion D
criterion_D(pop.size = c(25, 50, 200, 250, 500, 1000, 1500), 
            subcriteria = c("D"))
            
## Olny sub-criterion D2 (Please read the function details before its use)            
criterion_D(pop.size = NULL, 
            AOO = c(1,2,5,10,15,20,25), 
            n.Locs = c(1,1,2,3,4,5,6),
            subcriteria = c("D2"),
            AOO.threshold = 20,
            Loc.threshold = 5)
            
}
\references{
IUCN 2019. Guidelines for Using the IUCN Red List Categories and
Criteria. Version 14. Standards and Petitions Committee. Downloadable from:
http://www.iucnredlist.org/documents/RedListGuidelines.pdf.
}
\author{
Renato A. Ferreira de Lima & Gilles Dauby
}
