% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_text.R
\name{decode_text}
\alias{decode_text}
\title{Decode Text for Aesthetics}
\usage{
decode_text(x, ...)
}
\arguments{
\item{x}{A character vector.}

\item{...}{Dots.}
}
\value{
A visually-appealing character vector.
}
\description{
Transform text which is used for code to something more normal-looking.
}
\examples{
colnames(fake_health_ec_data)
colnames(fake_health_ec_data) |> decode_text()
}
