% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNeval.functionv11.R
\name{.EOO.comp}
\alias{.EOO.comp}
\title{Internal function}
\usage{
.EOO.comp(
  XY,
  exclude.area = FALSE,
  buff_width = 0.1,
  country_map = NULL,
  Name_Sp = "tax",
  alpha.hull = FALSE,
  convex.hull = TRUE,
  alpha = 1,
  buff.alpha = 0.1,
  method.less.than3 = "not comp"
)
}
\arguments{
\item{XY}{data.frame}

\item{exclude.area}{logical}

\item{buff_width}{numeric}

\item{country_map}{SpatialPolygonDataframe}

\item{Name_Sp}{string}

\item{alpha.hull}{logical}

\item{convex.hull}{logical}

\item{alpha}{integer}

\item{buff.alpha}{numeric}

\item{method.less.than3}{string}
}
\description{
EOO estimation
}
