% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_dist.R
\name{dummy_dist}
\alias{dummy_dist}
\title{Generate dummy distribution data}
\usage{
dummy_dist(
  nsp = 10,
  max_occ = 10,
  xmin = 0,
  xmax = 5,
  ymin = 0,
  ymax = 5,
  step = 0.1
)
}
\arguments{
\item{nsp}{integer}

\item{max_occ}{integer}

\item{xmin}{numeric}

\item{xmax}{numeric}

\item{ymin}{numeric}

\item{ymax}{numeric}

\item{step}{numeric}
}
\value{
A data.frame with three columns
}
\description{
Generates dummy geographic distribution data for multiple species
}
\examples{
dummy_dist(nsp = 5) ## five species dataset
dummy_dist(xmin = -20, xmax = 30, nsp = 30)

}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
