% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOO.comp.R
\name{EOO.comp}
\alias{EOO.comp}
\title{Compute Extent of Occurrence}
\usage{
EOO.comp(
  XY,
  method.range = "convex.hull",
  alpha = 1,
  buff.alpha = 0.1,
  method.less.than3 = "not comp",
  mode = "spheroid",
  proj_type = NULL,
  reproject = FALSE
)
}
\arguments{
\item{XY}{data.frame}

\item{method.range}{string, by default "convex.hull", can also take
"alpha.hull"}

\item{alpha}{integer}

\item{buff.alpha}{numeric}

\item{method.less.than3}{string}

\item{mode}{character string either 'spheroid' or 'planar'. By default
'spheroid'}

\item{proj_type}{crs}

\item{reproject}{logical FALSE whether the polygon should be converted to geographic coordinates if mode is \code{planar}}

\item{exclude.area}{logical. Default if FALSE}

\item{country_map}{SpatialPolygonDataframe. Default if NULL}
}
\value{
A list
\enumerate{
\item EOO a numeric vector of AOO estimates for each taxa
\item spatial.polygon a simple feature collection
}
}
\description{
Compute EOO given a data frame of coordinate in decimals degrees
}
\author{
Gilles Dauby & Renato A. Ferreira de Lima
}
\keyword{internal}
