% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.ex.R
\name{example_tutorial}
\alias{example_tutorial}
\title{Dataset to run the package introductory vignette}
\format{
A data frame with 1037 rows and 6 columns
}
\source{
\url{https://doi.org/10.15468/dl.mzmat2}
}
\usage{
data(example_tutorial)
}
\description{
A dataset containing occurrence records from three arborescent species from
the Atlantic Forest obtained from different sources, but mostly from GBIF.
}
\details{
Only records considered to have a high confidence level in terms of
georefferencing and species identifications are included.

Columns are: ddlat, ddlon, tax, coly, UC, source
}
\keyword{datasets}
