% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_docx.R
\name{format_docx}
\alias{format_docx}
\title{Produce a Word Document}
\usage{
format_docx(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Supply any arguments to
\code{\link[officedown]{rdocx_document}}. This allows arguments to be
supplied in the YAML in the same way as with any standard format, just
remember to use spaces to indent, not tabs.}
}
\value{
\code{officedown::rdocx_document} to pass to \code{rmarkdown::render}.
}
\description{
Provides a reference document with styles.
}
