% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_html.R
\name{format_html}
\alias{format_html}
\title{Produce an HTML Report}
\usage{
format_html()
}
\value{
\code{rmarkdown::html_document}
}
\description{
Produce an .html file from .Rmd, with branding and some preferences.
}
