% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.patches.R
\name{get.patches}
\alias{get.patches}
\title{Internal function}
\usage{
get.patches(points, Cell_size_AOO = 2, EOO.poly, proj_user, Resol_sub_pop = 5)
}
\arguments{
\item{points}{XY data frame}

\item{Cell_size_AOO}{numeric}

\item{EOO.poly}{Spatial polygon}

\item{proj_user}{projected coordinate system (in meters)}

\item{Resol_sub_pop}{numeric. Defines in kilometers the radius of the circles around each occurrence}
}
\description{
Get Occupied Cells and Patches
}
\details{
The ...
}
\examples{

mydf <- data.frame(ddlat = c(-44.6,-46.2,-45.4,-42.2,-43.7,-45.0,-28.0),
                   ddlon = c(-42.2,-42.6,-45.3,-42.5,-42.3,-39.0,-17.2),
                   tax = rep("a", 7),
                   valid = c(c(TRUE,TRUE,TRUE,FALSE,TRUE,TRUE,FALSE)),
                   stringsAsFactors = FALSE)
mydf$classes = as.double(mydf$valid)
shp <- SpatialPolygonsDataFrame(EOO.computing(mydf[mydf$valid,], export_shp = TRUE)[[2]], data.frame(tax = "a"))
plot(mydf[,2:1])
plot(shp, add=TRUE)
.over.valid.poly(shp, mydf)
.over.valid.poly(shp, mydf, proj_user = 5641)
.over.valid.poly(shp, mydf, proj_user = 5641, value = "flag")  

}
