% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.patches.R
\name{get.patches}
\alias{get.patches}
\title{Internal function}
\usage{
get.patches(
  points,
  cell_size = 2,
  nbe_rep = 0,
  EOO.poly = NULL,
  proj_type = "3857",
  Resol_sub_pop = 50
)
}
\arguments{
\item{points}{XY data frame}

\item{cell_size}{numeric}

\item{nbe_rep}{numeric}

\item{EOO.poly}{Spatial polygon}

\item{proj_type}{projected coordinate system (in meters)}

\item{Resol_sub_pop}{numeric. Defines in kilometers the radius of the circles around each occurrence}
}
\description{
Get Occupied Cells and Patches
}
\details{
The ...

"A taxon can be considered to be severely fragmented if most (>50%) of its total area of occupancy 
is in habitat patches that are (...) (2) separated from other habitat patches by a large distance."
}
\examples{

mydf <- data.frame(ddlat = c(-44.6,-46.2,-45.4,-42.2,-43.7,-45.0,-48.0),
                   ddlon = c(-42.2,-42.6,-45.3,-42.5,-42.3,-39.0,-37.2),
                   tax = rep("a", 7),
                   stringsAsFactors = FALSE)

}
