% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_df.R
\name{knit_df}
\alias{knit_df}
\title{Knit nice dataframes in .docx and .html}
\usage{
knit_df(...)
}
\arguments{
\item{...}{Unused dots.}
}
\value{
\code{\link[base]{.S3method}}.
}
\description{
Redefine the default print method for objects with class
\code{data.frame} when the rmarkdown file is knitted. If the file is knitted to
html, then \code{data.frame} objects will be output as
\code{\link[DT]{datatable}}. Knitting to Word (.docx) produces a
\code{\link[flextable]{flextable}}.

For further info, see \code{\link[knitr]{knit_print}} and
\code{\link[base]{.S3method}} and
\url{https://github.com/rstudio/rmarkdown-cookbook/issues/186}.

If you need to modify the method, define \code{knit_print.data.frame} in the
document.

\code{\link[reactable]{reactable}} is another alternative to
\code{\link[DT]{datatable}}.

It may be better to use the code underneath this function for transparency
and remove this function. The rmarkdown file just needs
\code{.S3method("knit_print", "data.frame", fn_name)} where fn_name could be
knit_docx_df, knit_html_df, wrap_reactable, or a function defined in the
document.
}
\seealso{
\code{\link[=knit_docx_df]{knit_docx_df()}}, \code{\link[=knit_html_df]{knit_html_df()}}
}
