% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_df.R
\name{knit_df}
\alias{knit_df}
\title{Knit nice dataframes in .docx and .html}
\usage{
knit_df(...)
}
\arguments{
\item{...}{Usual dots.}
}
\value{
Probably nothing.
}
\description{
Redefine the default print method for objects with class
"data.frame". For further info:
https://github.com/rstudio/rmarkdown-cookbook/issues/186. If you need to
modify the method, define knit_print.data.frame in the document. I tried to
create two functions, such that I could pass arguments to the knitting fn
(changing DT_opts) but it seems too difficult.

Why DT over gt? DT: pages, choose pageLength, copy/csv buttons. gt: has
most of this here https://gt.rstudio.com/reference/opt_interactive.html,
but not the download option.

Currently not ideal as datatable() trims trailing zeros, leading to
inconsistent number of decimal places in a given column. DT::formatRound is
not an option as trailing zeros are never trimmed.
}
