% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_df.R
\name{knit_docx_df}
\alias{knit_docx_df}
\title{Knit method for dataframes in docx output}
\usage{
knit_docx_df(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{dots}
}
\value{
Returns the dataframe in the form of
\code{\link[flextable]{flextable}}, via \code{knitr::knit_print}.
}
\description{
Knit method for dataframes in docx output
}
\seealso{
\code{\link[=knit_df]{knit_df()}}, \code{\link[=knit_html_df]{knit_html_df()}}
}
