% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_df.R
\name{knit_html_df}
\alias{knit_html_df}
\title{Knit method for dataframes in html output}
\usage{
knit_html_df(
  df,
  DT_opts = list(rownames = TRUE, dom = "ltipB", filter = "top", scrollX = TRUE,
    pageLength = 5, lengthMenu = c(5, 10, 20), buttons = c("copy", "csv")),
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{DT_opts}{passed to DT}

\item{...}{dots}
}
\value{
Returns the dataframe in the form of \code{\link[DT]{datatable}}, via
\code{knitr::knit_print}.
}
\description{
Knit method for dataframes in html output
}
\seealso{
\code{\link[=knit_df]{knit_df()}}, \code{\link[=knit_docx_df]{knit_docx_df()}}
}
