% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{land}
\alias{land}
\title{land}
\format{
A dataframe
}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}
\usage{
data(land)
}
\description{
A SpatialPolygonDataFrame showing land cover
}
\keyword{datasets}
