% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.comp.R
\name{locations.comp}
\alias{locations.comp}
\title{Estimate the number of locations.}
\usage{
locations.comp(
  XY,
  method = "fixed_grid",
  nbe_rep = 0,
  threat_list = NULL,
  names_threat = NULL,
  threat_weight = NULL,
  cell_size_locations = 10,
  method_polygons = c("no_more_than_one"),
  id_shape = "id_orig",
  Rel_cell_size = 0.05,
  parallel = FALSE,
  NbeCores = 2,
  show_progress = TRUE,
  proj_type = "cea"
)
}
\arguments{
\item{XY}{\link[base:data.frame]{data.frame}, see details}

\item{method}{string, indicating the method used for estimating the number of locations. See Details
\itemize{
\item \code{"fixed_grid"} (the default)
\item \code{"sliding_scale"}
}}

\item{nbe_rep}{numeric, the number of raster with random starting position for estimating the number of locations By default, it is 0 but some minimal translation of the raster are still done}

\item{threat_list}{list or sfc objects POLYGON or MULTIPOLYGON documenting threats. If provided, this will be taken into account for calculating number of location (see Details and \code{method_polygons}). By default, no shapefile is provided}

\item{names_threat}{character vector, indicating names of threats, optional}

\item{threat_weight}{numeric vector, indicating weight given to each threat}

\item{cell_size_locations}{numeric, value indicating the grid size in kilometres used for estimating the number of location. By default, equal to 10}

\item{method_polygons}{string. Used if \code{threat_list} is provided. See Details.
\itemize{
\item \code{"no_more_than_one"} (the default): each single POLYGON will be considered as a single location
\item \code{"grid"}: a grid of \code{cell_size_locations} size will be used to estimate the number of location within polygons
}}

\item{id_shape}{string}

\item{Rel_cell_size}{numeric, if \code{method_locations="sliding_scale"}, \code{cell_size_locations} is ignored
and the resolution is given by the maximum distance separating two occurrences multiplied by \code{Rel_cell_size}. By default, it is 0.05}

\item{parallel}{logical, whether running in parallel. By default, it is FALSE}

\item{NbeCores}{integer, register the number of cores for parallel execution. By default, it is 2}

\item{show_progress}{logical, whether a bar showing progress in computation should be shown. By default, it is TRUE}

\item{proj_type}{string or numeric}
}
\value{
A list
\enumerate{
\item a \code{data.frame} of two columns with the number of locations and potential issue for each species
\item a \code{sf} object representing the squared polygons OR a list of \code{sf} if \code{threat_list} is TRUE
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Estimate the number of locations (sensu IUCN) for multiple taxa,
taking into account spatial threats if provided.
}
\details{
\subsection{}{

\code{XY} as a \link[base:data.frame]{data.frame} should have the following structure:

\strong{It is mandatory to respect field positions, but field names do not matter}

\enumerate{
\item The first column is contains numeric value i.e. latitude in decimal degrees
\item The second column is contains numeric value i.e. longitude in decimal degrees
\item The third column is contains character value i.e. the names of the species
}

Locations are estimated by overlaying a grid of a given resolution (see \code{cell_size_locations} for
specifying the resolution). The number of locations is  the number of
occupied locations. Note that the grid position is overlaid in order to
minimize the number of locations (several translation of the grid are
performed and the one providing the minimum number of occupied cells is
provided).

If \code{threat_list} is provided,

which means occurrences within polygon documenting threats (if provided) will not be taken into account for estimating the number of locations following the grid system,

If \code{method} is "fixed_grid" as it is by default, the resolution is fixed and determined
by the argument \code{cell_size_locations}.
If \code{method} is "sliding_scale", the resolution is defind as 1/x*max.dist where max.dist is the maximum distance between any pairs of occurrences
and x is a defined parameter. 1/x is defined by \code{Rel_cell_size} argument and is 0.05 by default.
See Rivers M.C. et al. (2010) for more information on the methods.
}
}
\examples{
data(dataset.ex)
\dontrun{
locations <- locations.comp(dataset.ex)
}

# This would estimate the number of locations for all taxa by overlaying 
# randomly a grid 100 times. For each taxa, the minimum value is kept
\dontrun{
locations <- locations.comp(dataset.ex, nbe_rep = 100)
}

}
\references{
Gaston & Fuller 2009 The sizes of species'geographic ranges, Journal of Applied Ecology, 49 1-9
}
\author{
Gilles Dauby, \email{gildauby@gmail.com}
}
