% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Madagascar.protec.R
\name{Madagascar.protec}
\alias{Madagascar.protec}
\title{Dataset of Malagasy protected areas}
\format{
A SpatialPolygonDataFrame
}
\source{
World Database on Protected Areas http://protectedplanet.net/country/MG
}
\usage{
data(Madagascar.protec)
}
\description{
A SpatialPolygonDataFrame showing Malagasy protected areas
}
\references{
http://protectedplanet.net/country/MG
}
\keyword{datasets}
