% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop.comp.R
\name{subpop.comp}
\alias{subpop.comp}
\title{Estimate the number of subpopulations for one or multiple species}
\usage{
subpop.comp(
  XY,
  resol_sub_pop = NULL,
  proj_type = "cea",
  export_shp = FALSE,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2
)
}
\arguments{
\item{XY}{a data frame containing the geographical coordinates for each taxon
(see Details).}

\item{resol_sub_pop}{a value defining the radius of the circles around each
occurrence (in kilometres) or data frame vector containing a column 'tax'
with the taxa names and a column 'radius' with the species-specific radius
(in kilometre as well). Typically, this data frame is the output of
\code{ConR} function \code{subpop.radius}.}

\item{proj_type}{string or numeric}

\item{export_shp}{logical. Whether the resulting shapefiles should be
exported. FALSE by default.}

\item{parallel}{a logical. Whether running should be performed in parallel.
FALSE by default.}

\item{show_progress}{logical. Whether progress informations should displayed. TRUE by default}

\item{NbeCores}{an integer. Register the number of cores for parallel
execution. Two by default.}
}
\value{
If \code{export_shp} is TRUE,
\enumerate{
\item number_subpop a numeric vector of AOO estimates for each taxa
\item poly_subpop a simple feature collection
}

a \verb{Simple feature collection} with as many MULTIPOLYGON as taxa.
If \code{export_shp} is FALSE, a vector with estimated number of subpopulation per
taxa.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Estimate the number of subpopulations following the method
\strong{circular buffer method} (overlapping buffered circles form a single
subpopulation)
}
\details{
\code{XY} as a \link[base:data.frame]{data.frame} should have the following structure:

\strong{It is mandatory to respect field positions, but field names do not matter}

\enumerate{
\item The first column is contains numeric value i.e. latitude in decimal degrees
\item The second column is contains numeric value i.e. longitude in decimal degrees
\item The third column is contains character value i.e. the names of the species
}
}
\examples{
data(dataset.ex)

subpop.comp(dataset.ex, resol_sub_pop = 5)
rad.df <- data.frame(
    tax = unique(dataset.ex$tax),
    radius = seq(3,13, by=2),
    stringsAsFactors = FALSE
  )
subpop.comp(dataset.ex, resol_sub_pop = rad.df)
subpop.comp(dataset.ex, resol_sub_pop = rad.df, export_shp = TRUE)

}
\references{
Rivers MC, Bachman SP, Meagher TR, Lughadha EN, Brummitt NA
(2010) Subpopulations, locations and fragmentation: applying IUCN red list
criteria to herbarium specimen data. Biodiversity and Conservation 19:
2071-2085. doi: 10.1007/s10531-010-9826-9
}
\author{
Gilles Dauby & Renato A. Ferreira de Lima
}
