% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop.comp.R
\name{subpop.comp}
\alias{subpop.comp}
\title{Number of Subpopulations}
\usage{
subpop.comp(
  XY,
  Resol_sub_pop = NULL,
  proj_type = "cea",
  export_shp = FALSE,
  parallel = FALSE,
  show_progress = TRUE,
  NbeCores = 2
)
}
\arguments{
\item{XY}{a data frame containing the geographical coordinates for each taxon
(see Details).}

\item{Resol_sub_pop}{a value defining the radius of the circles around each
occurrence (in kilometres) or data frame vector containing a column 'tax' 
with the taxa names and a column 'radius' with the species-specific radius
(in kilometre as well). Typically, this data frame is the output of
```ConR``` function ```subpop.radius```.}

\item{proj_type}{character string or numeric or object of CRS class, by
default is "cea"}

\item{export_shp}{logical. Whether the resulting shapefiles should be
exported. FALSE by default.}

\item{parallel}{logical. Whether compute should run in parallel. FALSE by default.}

\item{show_progress}{logical. Whether a bar showing progress in computation
should be shown. TRUE by default.}

\item{NbeCores}{integer. Number of cores for parallel computation. Default to
2.}
}
\value{
If `export_shp` is TRUE, a list with [[1]]number_subpop and [[2]]poly_subpop a `Simple feature collection` with as many MULTIPOLYGON as taxa
If `export_shp` is FALSE, a vector with estimated number of subpopulation per taxa
}
\description{
Estimate the number of subpopulations following the method
  **circular buffer method** (overlapping buffered circles form a single subpopulation)
}
\details{
\strong{Input} as a \code{dataframe} should have the following structure:

\strong{It is mandatory to respect field positions, but field names do not matter}

\tabular{ccc}{
  [,1] \tab ddlat \tab numeric, latitude (in decimal degrees)\cr
  [,2] \tab ddlon \tab numeric, longitude (in decimal degrees)\cr
  [,3] \tab tax \tab character or factor, taxa names\cr
}
}
\examples{
data(dataset.ex)

subpop.comp(dataset.ex, Resol_sub_pop = 5)
rad.df <- data.frame(
    tax = unique(dataset.ex$tax),
    radius = seq(3,13, by=2),
    stringsAsFactors = FALSE
  )
subpop.comp(dataset.ex, Resol_sub_pop = rad.df)
subpop.comp(dataset.ex, Resol_sub_pop = rad.df, export_shp = TRUE)

}
\references{
Rivers MC, Bachman SP, Meagher TR, Lughadha EN, Brummitt NA
  (2010) Subpopulations, locations and fragmentation: applying IUCN red list
  criteria to herbarium specimen data. Biodiversity and Conservation 19:
  2071-2085. doi: 10.1007/s10531-010-9826-9
}
\author{
Gilles Dauby & Renato A. Ferreira de Lima
}
