% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop.estimation.R
\name{subpop.estimation}
\alias{subpop.estimation}
\title{Internal function}
\usage{
subpop.estimation(XY, Resol_sub_pop, export_shp = FALSE, proj_type)
}
\arguments{
\item{XY}{data.frame with first two columns are projected coordinates (numeric)}

\item{Resol_sub_pop}{numeric. Defines the radius of the circles
around each occurrence, in kilometres.}

\item{export_shp}{logical}

\item{proj_type}{}
}
\description{
Subpopulations estimation
}
