% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_and_date.R
\name{write_and_date}
\alias{write_and_date}
\title{Save Output File with the Date}
\usage{
write_and_date(input, ...)
}
\arguments{
\item{input}{File to be knitted.}

\item{...}{}
}
\value{
Output file created.
}
\description{
Save Output File with the Date
}
